S4 class BM.code

A class for storing and accessing up to 256 values using a RAW "big.matrix".

as.BM.code(x, code)

# S4 method for big.matrix,numeric
as.BM.code(x, code)

# S4 method for big.matrix,missing
as.BM.code(x, code)

# S4 method for big.matrix.descriptor,numeric
as.BM.code(x, code)

# S4 method for big.matrix.descriptor,missing
as.BM.code(x, code)

Arguments

x

Either a big.matrix or a big.matrix.descriptor.

code

A numeric vector of length 256. You should contruct it with rep(NA_real_, 256) and then replace the values which are of interest for you.

Details

Compared to a big.matrix, it adds a slot code which is used as a lookup table of size 256.

Examples

X <- big.matrix(10, 10, type = "raw")
X[] <- sample(as.raw(0:3), size = length(X), replace = TRUE)
X[,]

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   01   02   01   02   03   01   00   02   02    01
#>  [2,]   01   00   02   00   03   00   01   03   00    01
#>  [3,]   01   00   02   02   01   00   01   02   01    01
#>  [4,]   03   02   01   01   03   00   02   00   02    01
#>  [5,]   02   01   03   00   03   00   00   02   01    00
#>  [6,]   00   03   03   02   03   02   02   00   03    01
#>  [7,]   02   03   00   02   00   03   02   01   00    03
#>  [8,]   01   02   01   03   02   02   02   02   01    02
#>  [9,]   02   01   01   03   00   00   01   01   03    01
#> [10,]   03   00   03   00   00   03   03   01   00    00

code <- rep(NA_real_, 256)
code[1:3] <- c(1, 3, 5)

X.code <- as.BM.code(X, code)
X.code[,]

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    3    5    3    5   NA    3    1    5    5     3
#>  [2,]    3    1    5    1   NA    1    3   NA    1     3
#>  [3,]    3    1    5    5    3    1    3    5    3     3
#>  [4,]   NA    5    3    3   NA    1    5    1    5     3
#>  [5,]    5    3   NA    1   NA    1    1    5    3     1
#>  [6,]    1   NA   NA    5   NA    5    5    1   NA     3
#>  [7,]    5   NA    1    5    1   NA    5    3    1    NA
#>  [8,]    3    5    3   NA    5    5    5    5    3     5
#>  [9,]    5    3    3   NA    1    1    3    3   NA     3
#> [10,]   NA    1   NA    1    1   NA   NA    3    1     1