Linear kernel matrices for a big.matrix.

Compute linear kernel matrices for a "big.matrix" after applying a particular scaling to it.

BigXYt(X, block.size, ind.train = seq(nrow(X)), vec.center = rep(0,
  ncol(X)), vec.scale = rep(1, ncol(X)), use.Eigen = TRUE,
  progress = TRUE)

Arguments

X

A big.matrix. You shouldn't have missing values in your data.

block.size

Maximum number of columns read at once.

ind.train

An optional vector of the row indices that are used, for the training part. If not specified, all data are used.

vec.center

Vector that will be subtracted to the matrix, columnwise. Typically, the mean of each column. See colmeans.

vec.scale

Vector that will be divided to the matrix (after the substraction), columnwise. Typically, the sd of each column. See colsds.

use.Eigen

Use the Eigen library to compute \(X X^T\)? TRUE is the default. If FALSE, use R's tcrossprod. See details.

progress

Use a progress bar for the computation of the correlation matrix? Default is TRUE.

Value

Either

A big.matrix of type double if all rows are used in ind.train.

Two big.matrix of type double. One for \(X.train X.train^T\) to get Principal Components and one for \(X.test X.train^T\) to project the rest of the data.

Details

To compute \(X X^T\), using Eigen library is faster. However, if you link R with an optimized math library, using R's tcrossprod can be faster.

For example, you can easily link R with the https://software.intel.com/en-us/intel-mkl (Intel® MKL) through https://mran.revolutionanalytics.com/open/ (MRO). It really improves performance of R and RcppArmadillo matrix computations, yet not the ones of RcppEigen (at least not directly).

So,

Eigen should be prefered if you don't change anything,

base R should be prefered if you use MRO,

Eigen may be prefered if you manage to link RcppEigen with the MKL (please mailto:florian.prive.21@gmail.com if you do!).

See also

tcrossprod

Examples

# Simulating some data
X <- big.matrix(19, 43, shared = FALSE)
X[] <- rnorm(length(X))


# center and scale
vec.center <- colmeans(X)
vec.scale <- colsds(X)

# Comparing with tcrossprod
test <- BigXYt(X = X,
               block.size = 10,
               vec.center = vec.center,
               vec.scale = vec.scale)

#>   |                                                                              |                                                                      |   0%  |                                                                              |==============                                                        |  20%  |                                                                              |============================                                          |  40%  |                                                                              |==========================================                            |  60%  |                                                                              |========================================================              |  80%  |                                                                              |======================================================================| 100%


mat <- sweep(sweep(X[,], 2, vec.center, '-'), 2, vec.scale, '/')
diff <- test[,] - tcrossprod(mat)
print(max(abs(diff)))

#> [1] 2.131628e-14



# Using only half of the data for "training"
ind <- sort(sample(nrow(X), nrow(X)/2))
vec.center <- colmeans(X, ind)
vec.scale <- colsds(X, ind)

test <- BigXYt(X = X,
               block.size = 10,
               ind.train = ind,
               vec.center = vec.center,
               vec.scale = vec.scale)

#>   |                                                                              |                                                                      |   0%  |                                                                              |==============                                                        |  20%  |                                                                              |============================                                          |  40%  |                                                                              |==========================================                            |  60%  |                                                                              |========================================================              |  80%  |                                                                              |======================================================================| 100%


mat1 <- sweep(sweep(X[ind, ],  2, vec.center, '-'), 2, vec.scale, '/')
mat2 <- sweep(sweep(X[-ind, ], 2, vec.center, '-'), 2, vec.scale, '/')
diff1 <- test[[1]][,] - tcrossprod(mat1)
print(max(abs(diff1)))

#> [1] 1.065814e-14

diff2 <- test[[2]][,] - tcrossprod(mat2, mat1)
print(max(abs(diff2)))

#> [1] 7.105427e-15