Crossprod for a big.matrix.

Compute \(X^T X\) for a "big.matrix" X after applying a particular scaling to it.

BigXtX(X, block.size, ind.train = seq(nrow(X)), vec.center = rep(0,
  ncol(X)), vec.scale = rep(1, ncol(X)), use.Eigen = TRUE,
  progress = TRUE)

Arguments

X

A big.matrix. You shouldn't have missing values in your data.

block.size

Maximum number of columns read at once.

ind.train

An optional vector of the row indices that are used, for the training part. If not specified, all data are used.

vec.center

Vector that will be subtracted to the matrix, columnwise. Typically, the mean of each column. See colmeans.

vec.scale

Vector that will be divided to the matrix (after the substraction), columnwise. Typically, the sd of each column. See colsds.

use.Eigen

Use the Eigen library to compute \(X^T X\)? TRUE is the default.

progress

Use a progress bar for the computation of the correlation matrix? Default is TRUE.

Value

\(X.train^T X.train\)

See also

crossprod

Examples

# Simulating some data
X <- big.matrix(17, 41, shared = FALSE)
X[] <- rnorm(length(X))


# center and scale
vec.center <- colmeans(X)
vec.scale <- colsds(X)

# Comparing with tcrossprod
test <- BigXtX(X = X,
               block.size = 10,
               vec.center = vec.center,
               vec.scale = vec.scale)

#>   |                                                                              |                                                                      |   0%  |                                                                              |=====                                                                 |   7%  |                                                                              |=========                                                             |  13%  |                                                                              |==============                                                        |  20%  |                                                                              |===================                                                   |  27%  |                                                                              |=======================                                               |  33%  |                                                                              |============================                                          |  40%  |                                                                              |=================================                                     |  47%  |                                                                              |=====================================                                 |  53%  |                                                                              |==========================================                            |  60%  |                                                                              |===============================================                       |  67%  |                                                                              |===================================================                   |  73%  |                                                                              |========================================================              |  80%  |                                                                              |=============================================================         |  87%  |                                                                              |=================================================================     |  93%  |                                                                              |======================================================================| 100%


mat <- sweep(sweep(X[,], 2, vec.center, '-'), 2, vec.scale, '/')
diff <- test[,] - crossprod(mat)
print(max(abs(diff)))

#> [1] 0



# Using only half of the data for "training"
ind <- sort(sample(nrow(X), nrow(X)/2))
vec.center <- colmeans(X, ind)
vec.scale <- colsds(X, ind)

test <- BigXtX(X = X,
               block.size = 10,
               ind.train = ind,
               vec.center = vec.center,
               vec.scale = vec.scale)

#>   |                                                                              |                                                                      |   0%  |                                                                              |=====                                                                 |   7%  |                                                                              |=========                                                             |  13%  |                                                                              |==============                                                        |  20%  |                                                                              |===================                                                   |  27%  |                                                                              |=======================                                               |  33%  |                                                                              |============================                                          |  40%  |                                                                              |=================================                                     |  47%  |                                                                              |=====================================                                 |  53%  |                                                                              |==========================================                            |  60%  |                                                                              |===============================================                       |  67%  |                                                                              |===================================================                   |  73%  |                                                                              |========================================================              |  80%  |                                                                              |=============================================================         |  87%  |                                                                              |=================================================================     |  93%  |                                                                              |======================================================================| 100%


mat <- sweep(sweep(X[ind, ],  2, vec.center, '-'), 2, vec.scale, '/')
diff <- test[,] - crossprod(mat)
print(max(abs(diff)))

#> [1] 0