Train one model

Train one model

COPY_biglasso_part(
  X,
  y.train,
  ind.train,
  ind.col,
  covar.train,
  family,
  lambda,
  center,
  scale,
  resid,
  alpha,
  eps,
  max.iter,
  dfmax,
  ind.val,
  covar.val,
  y.val,
  n.abort,
  nlam.min,
  base.train,
  base.val,
  pf
)

Value

A named list with following variables:

intercept

A vector of intercepts, corresponding to each lambda.

beta

The vector of coefficients that minimized the loss on the validation set.

iter

A vector of length nlambda containing the number of iterations until convergence at each value of lambda.

lambda

The sequence of regularization parameter values in the path.

alpha

Input parameter.

loss

A vector containing either the residual sum of squares (for linear models) or negative log-likelihood (for logistic models) of the fitted model at each value of lambda.

loss.val

A vector containing the loss for the corresponding validation set.

message

Reason the fitting has stopped.

nb_active

The number of active (non-zero) variables along the regularization path.

nb_candidate

The number of candidate variables (used in the gradient descent) along the regularization path.

ind.train

Indices of training set.