Fit sparse linear SVM with lasso or elasti-net regularization

Fit solution paths for sparse linear SVM regularized by lasso or elastic-net over a grid of values for the regularization parameter lambda.

COPY_sparseSVM(X, y.train, ind.train, covar.train, alpha = 1, gamma = 0.1,
  nlambda = 100, lambda.min = if (nrow(X) > ncol(X)) 0.01 else 0.05, lambda,
  screen = c("ASR", "SR", "none"), max.iter = 1000, eps = 1e-05,
  dfmax = p + 1, penalty.factor = NULL, message = FALSE)

Arguments

alpha
The elastic-net mixing parameter that controls the relative contribution from the lasso and the ridge penalty. It must be a number between 0 and 1. alpha=1 is the lasso penalty and alpha=0 the ridge penalty.

gamma
The tuning parameter for huberization smoothing of hinge loss. Default is 0.1.

nlambda
The number of lambda values. Default is 100.

lambda.min
The smallest value for lambda, as a fraction of lambda.max, the data derived entry value. Default is 0.01 if the number of observations is larger than the number of variables and 0.05 otherwise.

lambda
A user-specified sequence of lambda values. Typical usage is to leave blank and have the program automatically compute a lambda sequence based on nlambda and lambda.min. Specifying lambda overrides this. This argument should be used with care and supplied with a decreasing sequence instead of a single value.

screen
Screening rule to be applied at each lambda that discards variables for speed. Either "ASR" (default), "SR" or "none". "SR" stands for the strong rule, and "ASR" for the adaptive strong rule. Using "ASR" typically requires fewer iterations to converge than "SR", but the computing time are generally close. Note that the option "none" is used mainly for debugging, which may lead to much longer computing time.

max.iter
Maximum number of iterations. Default is 1000.

eps
Convergence threshold. The algorithms continue until the maximum change in the objective after any coefficient update is less than eps times the null deviance. Default is 1E-7.

dfmax
Upper bound for the number of nonzero coefficients. The algorithm exits and returns a partial path if dfmax is reached. Useful for very large dimensions.

penalty.factor
A numeric vector of length equal to the number of variables. Each component multiplies lambda to allow differential penalization. Can be 0 for some variables, in which case the variable is always in the model without penalization. Default is 1 for all variables.

message
If set to TRUE, sparseSVM will inform the user of its progress. Default is FALSE.

Value

A named list containing:

call

The call that produced this object.

intercept

A vector of intercepts, corresponding to each lambda.

beta

The fitted matrix of coefficients. The number of rows is equal to the number of coefficients, and the number of columns is equal to nlambda.

iter

A vector of length nlambda containing the number of iterations until convergence at each value of lambda.

saturated

A logical flag for whether the number of nonzero coefficients has reached dfmax.

lambda

The sequence of regularization parameter values in the path.

alpha

Input parameter.

gamma

Input parameter.

penalty.factor

Input parameter.

levels

Levels of the output class labels.

Details

The sequence of models indexed by the regularization parameter lambda is fitted using a semismooth Newton coordinate descent algorithm. The objective function is defined to be $$\frac{1}{n} \sum hingeLoss(y_i (x_i' w + b)) + \lambda\textrm{penalty}(w).$$ where $$hingeLoss(t) = max(0, 1-t)$$ and the intercept b is unpenalized.