Split-Apply-Combine

A Split-Apply-Combine strategy to apply common R functions to a big.matrix.

big_apply(X., a.FUN, a.combine, ncores = 1, block.size = 1000,
  ind = cols_along(X.), ...)

Arguments

X.
Either a big.matrix or a big.matrix.descriptor.

a.FUN
The function to be applied to each subset matrix. It must take a big.matrix as first argument and ind, a vector of indices, which are used to split the data. Example: if you want to apply a function to X[ind.row, ind.col], you may use X[ind.row, ind.col[ind]] in a.FUN.

a.combine
function that is used by foreach to process the tasks results as they generated. This can be specified as either a function or a non-empty character string naming the function. Specifying 'c' is useful for concatenating the results into a vector, for example. The values 'cbind' and 'rbind' can combine vectors into a matrix. The values '+' and '*' can be used to process numeric data. By default, the results are returned in a list.

ncores
Number of cores used. Default doesn't use parallelism.

block.size
Maximum number of columns read at once. Default is 1000.

ind
Initial vector of subsetting indices. Default is the vector of all column indices.

...
Extra arguments to be passed to a.FUN.

Value

The result of foreach.

Details

This function splits indices in parts, then apply a given function to each subset matrix and finally combine the results. If parallelization is used, this function splits indices in parts for parallelization, then split again them on each core, apply a given function to each part and finally combine the results (on each cluster and then from each cluster).

See also

big_parallelize

Examples

X.desc <- big_attachExtdata()

# get the means of each column
colMeans_sub <- function(X, ind) colMeans(X[, ind])
colmeans <- big_apply(X.desc, a.FUN = colMeans_sub, a.combine = 'c')

# get the norms of each column
colNorms_sub <- function(X, ind) sqrt(colSums(X[, ind]^2))
colnorms <- big_apply(X.desc, colNorms_sub, a.combine = 'c')

# get the sums of each row
# split along rows: need to change the "complete" `ind` parameter
rowsums <- big_apply(X.desc, a.FUN = function(X, ind) rowSums(X[ind, ]),
                     ind = rows_along(X.desc), a.combine = 'c',
                     block.size = 100)
# it is usually preferred to split along columns
# because `big.matrix` are stored by column.
rowsums2 <- big_apply(X.desc, a.FUN = function(X, ind) rowSums(X[, ind]),
                      a.combine = '+')

## Every extra parameter to `a.FUN` should be passed to `big_apply`
# get the crossproduct between X and a matrix A
# note that we don't explicitly pass `ind.col` to `a.FUN`
body(big_cprodMat)

#> {
#>     stopifnot(length(ind.row) == nrow(A))
#>     big_apply(X., a.FUN = function(x, ind, A, ind.row) {
#>         crossprod(x[ind.row, ind], A)
#>     }, a.combine = "rbind", block.size = block.size, ind = ind.col, 
#>         ncores = ncores2, A = A, ind.row = ind.row)
#> }

# get the product between X and a matrix B
# here, we must explicitly pass `ind.col` to `a.FUN`
# because the right matrix also needs to be subsetted.
body(big_prodMat)

#> {
#>     stopifnot(length(ind.col) == nrow(A))
#>     big_apply(X., a.FUN = function(x, ind, A, ind.row, ind.col) {
#>         x[ind.row, ind.col[ind]] %*% A[ind, ]
#>     }, a.combine = "+", block.size = block.size, ind = seq_along(ind.col), 
#>         ncores = ncores2, A = A, ind.row = ind.row, ind.col = ind.col)
#> }