Attach a "big.matrix.descriptor"

Attach a "big.matrix.descriptor"

big_attach(descriptorfile)

big_attachExtdata()

Arguments

descriptorfile

Value

A big.matrix.descriptor object.

Examples

tmpfile <- tempfile()
X.desc <- describe(
  big.matrix(10, 10, backingfile = basename(tmpfile),
             backingpath = dirname(tmpfile),
             descriptorfile = paste0(basename(tmpfile), ".desc"))
)
X.desc <- describe(test)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list("numeric"), structure(function (x) standardGeneric("describe"), generic = structure("describe", package = "bigmemory.sri"), package = "bigmemory.sri", group = list(), valueClass = character(0), signature = "x", default = `\001NULL\001`, skeleton = (function (x) stop("invalid call in method dispatch to 'describe' (no default method)",     domain = NA))(x), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function ‘describe’ for signature ‘"numeric"’

descriptorfile <- paste0(tmpfile, ".desc")
X.desc2 <- big_attach(descriptorfile)
all.equal(X.desc, X.desc2)

#> [1] TRUE