Standard univariate statistics

Standard univariate statistics for columns of a big.matrix. For now, the sum and var are implemented (the mean and sd can easily be deduced, see examples).

big_colstats(X, ind.train = seq_len(nrow(X)))

Arguments

X

A big.matrix. You shouldn't have missing values in your data.

ind.train

An optional vector of the row indices that are used, for the training part. If not specified, all data are used.

Value

List of two numeric vectors sum and var with the corresponding column statistics.

See also

colSums apply

Examples

# Simulating some data
X <- big.matrix(41, 7)
X[] <- rnorm(length(X))

# Check the results
print(test <- big_colstats(X))

#> $sum
#> [1] -0.003010285  7.649088519  1.350159144  4.335068110  5.829699799
#> [6] -3.901864366  2.526804528
#> 
#> $var
#> [1] 1.0949455 1.1210312 0.9183947 0.8801451 1.5262439 0.8653047 1.1147029
#> 

# Only with the first 10 rows
ind <- 1:10
print(test <- big_colstats(X, ind))

#> $sum
#> [1] -1.0311363  2.3823528  0.7462801  2.9074259  0.3666418  0.2296901 -2.7341583
#> 
#> $var
#> [1] 1.3675630 1.6796840 0.9265305 0.7511339 2.0708160 0.6593327 1.5149690
#> 

print(all.equal(test$sum, colSums(X[ind, ])))

#> [1] TRUE

print(all.equal(test$var, apply(X[ind, ], 2, var)))

#> [1] TRUE

# deduce mean and sd
# note that the are also implemented in big_scale()
means <- test$sum / length(ind) # if using all rows,
                                # divide by nrow(X) instead
print(all.equal(means, colMeans(X[ind, ])))

#> [1] TRUE

sds <- sqrt(test$var)
print(all.equal(sds, apply(X[ind, ], 2, sd)))

#> [1] TRUE

# NA in, NA out
X[2, 5] <- NA
print(big_colstats(X))

#> $sum
#> [1] -0.003010285  7.649088519  1.350159144  4.335068110           NA
#> [6] -3.901864366  2.526804528
#> 
#> $var
#> [1] 1.0949455 1.1210312 0.9183947 0.8801451        NA 0.8653047 1.1147029
#>