Cross-product with a vector

Cross-product between a Filebacked Big Matrix and a vector.

big_cprodVec(X, y.row, ind.row = rows_along(X),
  ind.col = cols_along(X))

Arguments

X
A FBM.

y.row
A vector of same size as ind.row.

ind.row
An optional vector of the row indices that are used. If not specified, all rows are used. Don't use negative indices.

ind.col
An optional vector of the column indices that are used. If not specified, all columns are used. Don't use negative indices.

Value

\(X^T \cdot y\).

Examples

X <- big_attachExtdata()
n <- nrow(X)
m <- ncol(X)
y <- rnorm(n)

test <- big_cprodVec(X, y)             # vector
true <- crossprod(X[], y)  # one-column matrix
all.equal(test, as.numeric(true))

#> [1] TRUE

# subsetting
ind.row <- sample(n, n/2)
ind.col <- sample(m, m/2)

tryCatch(test2 <- big_cprodVec(X, y, ind.row, ind.col),
         error = function(e) print(e))

#> <Rcpp::exception in cpMatVec4(X, y.row, ind.row, ind.col): Incompatibility between dimensions.>

# returns an error. You need to use the subset of y:
test2 <- big_cprodVec(X, y[ind.row], ind.row, ind.col)
true2 <- crossprod(X[ind.row, ind.col], y[ind.row])
all.equal(test2, as.numeric(true2))

#> [1] TRUE