Crossprod

Compute \(X.train^T X.train\) for a big.matrix X after applying a particular scaling to it.

big_crossprodSelf(X, fun.scaling, ind.train = seq(nrow(X)),
  block.size = 1000, use.Eigen = TRUE, ...)

Arguments

X

A big.matrix. You shouldn't have missing values in your data.

fun.scaling

A function that returns a named list of mean and sd for every column, to scale each of their elements such as followed: $$\frac{X_{i,j} - mean_j}{sd_j}$$.

ind.train

An optional vector of the row indices that are used, for the training part. If not specified, all data are used.

block.size

Maximum number of columns read at once. Default is 1000.

use.Eigen

Should the Eigen library be used for matrix computations? Default is TRUE. See details.

...

Arguments passed on to bigmemory::big.matrix

dimnames

a list of the row and column names; use with caution for large objects.

separated

use separated column organization of the data; see details.

backingfile

the root name for the file(s) for the cache of x.

backingpath

the path to the directory containing the file backing cache.

descriptorfile

the name of the file to hold the backingfile description, for subsequent use with attach.big.matrix; if NULL, the backingfile is used as the root part of the descriptor file name. The descriptor file is placed in the same directory as the backing files.

binarydescriptor

the flag to specify if the binary RDS format should be used for the backingfile description, for subsequent use with attach.big.matrix; if NULL of FALSE, the dput() file format is used.

shared

TRUE by default, and always TRUE if the big.matrix is file-backed. For a non-filebacked big.matrix, shared=FALSE uses non-shared memory, which can be more stable for large (say, >50% of RAM) objects. Shared memory allocation can sometimes fail in such cases due to exhausted shared-memory resources in the system.

Value

\(X.train^T X.train\) as a big.matrix. Its dimensions are both ncol(X) and its type is double.

Details

For matrix computations, using Eigen library is faster. However, if you link R with an optimized math library, using R's base operations is even much faster.

For example, you can easily link R with the https://software.intel.com/en-us/intel-mkl (Intel® MKL) through https://mran.revolutionanalytics.com/open/ (MRO). It really improves performance of R and RcppArmadillo matrix computations, yet not the ones of RcppEigen (at least not directly).

So,

Eigen should be prefered if you don't change anything,

base R should be prefered if you use MRO,

Eigen may be prefered if you manage to link RcppEigen with the MKL (please mailto:florian.prive.21@gmail.com if you do!).

See also

crossprod

Examples

# Simulating some data
X <- big.matrix(41, 17)
X[] <- rnorm(length(X))

# Comparing with tcrossprod
big_noscale <- big_scale(center = FALSE)
test <- big_crossprodSelf(X, fun.scaling = big_noscale)
print(dim(test))

#> [1] 17 17

print(all.equal(test[,], crossprod(X[,])))

#> [1] TRUE

# Using only half of the data for "training"
ind <- sort(sample(nrow(X), nrow(X)/2))
test2 <- big_crossprodSelf(X, fun.scaling = big_noscale,
                           ind.train = ind)
print(dim(test2))

#> [1] 17 17

print(all.equal(test2[,], crossprod(X[ind, ])))

#> [1] TRUE