Product with a matrix

Product between a "big.matrix" and a matrix.
big_prodMat(X., A, ind.row = rows_along(X.), ind.col = cols_along(X.),
  block.size = 1000, ncores2 = 1)

Arguments

X. 	Either a big.matrix or a big.matrix.descriptor.

A 	A matrix with length(ind.col) rows.

ind.row 	An optional vector of the row indices that are used. If not specified, all rows are used. Don't use negative indices.

ind.col 	An optional vector of the column indices that are used. If not specified, all columns are used. Don't use negative indices.

block.size 	Maximum number of columns read at once. Default is 1000.

ncores2 	Number of cores used. Default doesn't use parallelism. For this function, use only half of the cores you have.

Value
\(X \cdot A\).
Examples

X.desc <- big_attachExtdata()
n <- nrow(X.desc)
m <- ncol(X.desc)
A <- matrix(0, m, 10); A[] <- rnorm(length(A))

test <- big_prodMat(X.desc, A)
true <- attach.BM(X.desc)[,] %*% A
all.equal(test, true)

#> [1] TRUE

# subsetting
ind.row <- sample(n, n/2)
ind.col <- sample(m, m/2)

## Not run: test2 &lt;- big_prodMat(X.desc, A, ind.row, ind.col)
# returns an error. You need to use the subset of A:
test2 <- big_prodMat(X.desc, A[ind.col, ], ind.row, ind.col)
true2 <- attach.BM(X.desc)[ind.row, ind.col] %*% A[ind.col, ]
all.equal(test2, true2)

#> [1] TRUE