A randomized algorithm for SVD.

A randomized algorithm for SVD (or PCA) of a "big.matrix".

big_randomSVD(X, fun.scaling = fun_scale, ind.train = seq(nrow(X)),
  block.size = 1000, K = 10, I = 20, use.Eigen = TRUE, ncores = 1,
  backingpath = NULL)

Arguments

X

A big.matrix. You shouldn't have missing values in your data.

fun.scaling

A function that returns a named list of mean and sd for every column, to scale each of their elements such as followed: \(\frac{X_{i,j} - mean_j}{sd_j}\).

ind.train

An optional vector of the row indices that are used, for the training part. If not specified, all data are used.

block.size

Maximum number of columns read at once. Default is 1000.

K

Number of PCs to compute. This algorithm shouldn't be used to compute a lot of PCs. Default is 10.

I

The number of iterations of the algorithm. Default is 20.

use.Eigen

Should the Eigen library be used for matrix computations? Default is TRUE. See details.

ncores

Number or cores used. Default doesn't use parallelism.

backingpath

If X is filebacked and parallelism is used, the path where are stored the files that are backing X.

References

Rokhlin, V., Szlam, A., & Tygert, M. (2010). A Randomized Algorithm for Principal Component Analysis. SIAM Journal on Matrix Analysis and Applications, 31(3), 1100–1124. doi:10.1137/080736417

See also

big_funScaling prcomp svd

Examples

# Simulating some data
X <- big.matrix(1023, 511)
X[] <- rnorm(length(X))

# Comparing with prcomp
test <- big_randomSVD(X = X,
                      fun.scaling = big_scale,
                      block.size = 50,
                      ncores = 1)

#> (1)

#> Error in expr_fun(i): task 1 failed - "there is no package called ‘flock’"

pca <- prcomp(X[,], center = TRUE, scale. = TRUE)

# same scaling
print(all.equal(test$means, pca$center))

#> Error in all.equal(test$means, pca$center): objet 'test' introuvable

print(all.equal(test$sds, pca$scale))

#> Error in all.equal(test$sds, pca$scale): objet 'test' introuvable

# scores or loadings are the same or opposite
scores <- test$u %*% diag(test$d)

#> Error in eval(expr, envir, enclos): objet 'test' introuvable

plot(as.numeric(scores), as.numeric(pca$x[, 1:10]))

#> Error in plot(as.numeric(scores), as.numeric(pca$x[, 1:10])): objet 'scores' introuvable

plot(as.numeric(test$v), as.numeric(pca$rotation[, 1:10]))

#> Error in plot(as.numeric(test$v), as.numeric(pca$rotation[, 1:10])): objet 'test' introuvable

# Using only half of the data for "training"
ind <- sort(sample(nrow(X), nrow(X)/2))

test2 <- big_randomSVD(X = X, ind.train = ind,
                       fun.scaling = big_scale,
                       block.size = 10, ncores = 1)

#> (2)

#> Error in expr_fun(i): task 1 failed - "there is no package called ‘flock’"

pca2 <- prcomp(X[ind, ], center = TRUE, scale. = TRUE)

# same scaling
print(all.equal(test2$means, pca2$center))

#> Error in all.equal(test2$means, pca2$center): objet 'test2' introuvable

print(all.equal(test2$sds, pca2$scale))

#> Error in all.equal(test2$sds, pca2$scale): objet 'test2' introuvable

# scores or loadings are the same or opposite
scores2 <- test2$u %*% diag(test2$d)

#> Error in eval(expr, envir, enclos): objet 'test2' introuvable

plot(as.numeric(scores2), as.numeric(pca2$x[, 1:10]))

#> Error in plot(as.numeric(scores2), as.numeric(pca2$x[, 1:10])): objet 'scores2' introuvable

plot(as.numeric(test2$v), as.numeric(pca2$rotation[, 1:10]))

#> Error in plot(as.numeric(test2$v), as.numeric(pca2$rotation[, 1:10])): objet 'test2' introuvable