Read a file

Read a file as a Filebacked Big Matrix and a data frame of meta information. For a mini-tutorial, please see this vignette.

big_read(file, sep, header, confirmed = FALSE, verbose = TRUE,
  ind.skip = integer(0), ind.meta = integer(0), nlines = NULL,
  nlines.block = NULL, fun.con = function(f) file(f, open = "rt"),
  type = NULL, ...)

Arguments

file
The path to the file to be read.

sep
The field separator character.

header
A logical value indicating whether the file contains the names of the variables as its first line.

confirmed
A logical value indicating whether you don't want to pause to confirm that everything has been parsed like you wanted. Default is FALSE and will ask you to confirm during execution.

verbose
Be verbose? Default is TRUE.

ind.skip
An integer vector specifying which columns to skip.

ind.meta
An integer vector specifying which columns should be part of the meta information. Non-numeric columns will automatically be added to meta information because they can't be stored in an FBM.

nlines
Number of lines of the file (including possible header). Default automatically gets this number thanks to a function of David Gohel.

nlines.block
Number of lines to be read at once. Default automatically gets this number based on getOption("bigstatsr.block.sizeGB").

fun.con
Function to open a connection from a file.

type
Type of the resulting Filebacked Big Matrix. Default is automatically guessing between "integer" and "double".

...
Arguments passed on to FBM

backingfile

Path to the file storing the Big Matrix on disk. An extension ".bk" will be automatically added. Default stores in the temporary directory.

create_bk

Create a backingfile (the default) or use an existing one (which should be named by the backingfile parameter and have an extension ".bk"). For example, this could be used to convert a filebacked big.matrix from package bigmemory to a FBM.

save

Whether to save the result object in an ".rds" file alongside the backingfile. Default is FALSE.

Value

A list with 3 elements:

A Filebacked Big Matrix.

A vector of colnames.

A data frame of meta information (e.g. non-numeric variables such as rownames)