Read a file

Read a file as a big.matrix. For a mini-tutorial, please see the vignette.

big_readBM(file, file.nheader = 0, info.nelem = 0, split = " ",
  read.what = double(), read.transfo = identity,
  BM.type = typeof(read.what), transpose = FALSE, fun.createBM = BM())

Arguments

file
The path to the file to be read.

file.nheader
Number of lines to read at the beginning of the file, as a separate header information. Default is 0. Each line is read as one string. You may need to postprocess it with strsplit. See examples.

info.nelem
Number of elements of extra information to read at the beginning of each line. Default is 0.

split
The separator used in the file. Default is a space.

read.what
What type of elements to scan? Default is double(). You can also use character() or integer().

read.transfo
Function that transforms each line you read.

BM.type
Type of the resulting big.matrix. Default uses the type of read.what. This can be useful if you have only small integers or that the read.transfo function transforms the type of what is read.

transpose
Should the resulting big.matrix be transposed? Default is FALSE.

fun.createBM
A function that creates a big.matrix. It should have exactly 3 parameters nrow, ncol and type. See FBM.

Value

A big.matrix (or its descriptor) with two attributes header and info.