Some scaling functions

Some scaling functions for a "big.matrix" to be used as the fun.scaling parameter of some functions of this package.

big_scale(center = TRUE, scale = TRUE)

Arguments

center

A logical value: whether to return means or 0s.

scale

A logical value: whether to return sds or 1s. You can't use scale without using center.

Value

A new function that returns a named list of two vectors "mean" and "sd" which are as long as the number of columns of X.

See also

scale

Examples

# Simulating some data
X <- big.matrix(41, 17)
X[] <- rnorm(length(X))

# No scaling
big_noscale <- big_scale(center = FALSE, scale = FALSE)
print(class(big_noscale)) # big_scale returns a new function

#> [1] "function"

print(big_noscale(X))

#> $mean
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $sd
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 

big_noscale2 <- big_scale(center = FALSE)
print(big_noscale2(X)) # you can't scale without centering

#> $mean
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $sd
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 

# Centering
big_center <- big_scale(scale = FALSE)
print(big_center(X))

#> $mean
#>  [1] -0.019761686  0.083248876  0.084472447  0.186460622 -0.150471109
#>  [6] -0.124280394 -0.075561993  0.306725936  0.113871759 -0.038789795
#> [11]  0.007514213 -0.321323174  0.339138036 -0.243482975  0.129367918
#> [16] -0.016379115  0.249696304
#> 
#> $sd
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 

# + scaling
print(test <- big_scale()(X, 1:5)) # note that big_scale() is a function

#> $mean
#>  [1] -0.20953714  0.29286534 -0.62326664 -0.06150171  0.06306785 -0.59441531
#>  [7] -0.63106140 -0.06047521 -0.69039624  0.26607624  0.15526762  0.12189187
#> [13]  1.26826336  0.11655883  0.45904252 -0.36210139  0.14378307
#> 
#> $sd
#>  [1] 0.5600132 0.7671057 0.7859392 0.7370883 0.8602522 0.3980393 0.6227674
#>  [8] 1.1881167 1.6149365 1.0207927 0.5377332 0.7119933 0.9573899 0.9341371
#> [15] 1.2132874 0.9708425 1.3499851
#> 

print(all.equal(test$mean, colMeans(X[1:5, ])))

#> [1] TRUE

print(all.equal(test$sd, apply(X[1:5, ], 2, sd)))

#> [1] TRUE