Sparse linear regression

Fit lasso penalized linear regression path for a big.matrix. Covariates can be added to correct for confounders.

big_spLinReg(X, y.train, ind.train = seq(nrow(X)), covar.train = NULL, ...)

Arguments

X

A big.matrix. You shouldn't have missing values in your data.

y.train

Vector of responses, corresponding to ind.train.

ind.train

An optional vector of the row indices that are used, for the training part. If not specified, all rows are used.

covar.train

Matrix of covariables to be added in each model to correct for confounders (e.g. the scores of PCA), corresponding to ind.train. Default is NULL and corresponds to only adding an Intercept to each model.

...

Arguments passed on to COPY_biglasso

alpha

The elastic-net mixing parameter that controls the relative contribution from the lasso (l1) and the ridge (l2) penalty. The penalty is defined as $$ \alpha||\beta||_1 + (1-\alpha)/2||\beta||_2^2.$$ alpha = 1 is the lasso penalty and alpha in between 0 (1e-6) and 1 is the elastic-net penalty.

lambda.min

The smallest value for lambda, as a fraction of lambda.max. Default is .001 if the number of observations is larger than the number of covariates and .01 otherwise.

nlambda

The number of lambda values. Default is 100.

lambda.log.scale

Whether compute the grid values of lambda on log scale (default) or linear scale.

lambda

A user-specified sequence of lambda values. By default, a sequence of values of length nlambda is computed, equally spaced on the log scale.

eps

Convergence threshold for inner coordinate descent. The algorithm iterates until the maximum change in the objective after any coefficient update is less than eps times the null deviance. Default value is 1e-7.

max.iter

Maximum number of iterations. Default is 1000.

dfmax

Upper bound for the number of nonzero coefficients. Default is no upper bound. However, for large data sets, computational burden may be heavy for models with a large number of nonzero coefficients.

penalty.factor

A multiplicative factor for the penalty applied to each coefficient. If supplied, penalty.factor must be a numeric vector of length equal to sum of the number of columns of X and the number of covariables (intercept excluded). The purpose of penalty.factor is to apply differential penalization if some coefficients are thought to be more likely than others to be in the model. Current package doesn't allow unpenalized coefficients. That is penalty.factor cannot be 0.

warn

Return warning messages for failures to converge and model saturation? Default is TRUE.

verbose

Whether to print out the start, the timing of each lambda iteration and the end. Default is FALSE.

Value

A named list with following variables:

Details

This is a modified version of one function of https://github.com/YaohuiZeng/biglasso. It adds the possibility to train to models with covariables and use many types of big.matrix (not only double ones). Yet, it only corresponds to screen = "SSR".

References

Tibshirani, R., Bien, J., Friedman, J., Hastie, T., Simon, N., Taylor, J. and Tibshirani, R. J. (2012), Strong rules for discarding predictors in lasso-type problems. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 74: 245–266. http://dx.doi.org/10.1111/j.1467-9868.2011.01004.x.

Zeng, Y., and Breheny, P. (2016). The biglasso Package: A Memory- and Computation-Efficient Solver for Lasso Model Fitting with Big Data in R. arXiv preprint arXiv:1701.05936. https://arxiv.org/abs/1701.05936.

See also

glmnet biglasso

Examples

# simulating some data
N <- 73
M <- 430
X <- as.big.matrix(matrix(rnorm(N*M, sd = 5), N), type = "integer")

#> Warning: Assignment will down cast from double to integer
#> Hint: To remove this warning type:  options(bigmemory.typecast.warning=FALSE)

y <- rnorm(N)
covar <- matrix(rnorm(N * 3), N)

# error, only handle `double` `big.matrix` objects
tryCatch({
  test <- biglasso::biglasso(X, y, family = "gaussian")
  print(summary(test$scale))
}, error = function(e) message("One error has been catched."))

#> One error has been catched.

# OK here
test2 <- big_spLinReg(X, y)
str(test2)

#> List of 16
#>  $ intercept     : Named num [1:100] 0.0192 0.0205 0.0228 0.0273 0.031 ...
#>   ..- attr(*, "names")= chr [1:100] "0.2889" "0.2758" "0.2632" "0.2513" ...
#>  $ beta          :Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   .. ..@ i       : int [1:5198] 74 74 399 74 344 399 74 221 344 399 ...
#>   .. ..@ p       : int [1:101] 0 0 1 3 6 11 17 25 33 43 ...
#>   .. ..@ Dim     : int [1:2] 430 100
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:100] "0.2889" "0.2758" "0.2632" "0.2513" ...
#>   .. ..@ x       : num [1:5198] 0.00313 0.00603 -0.00116 0.00882 -0.00147 ...
#>   .. ..@ factors : list()
#>  $ iter          : int [1:100] 0 3 5 6 5 6 6 4 6 8 ...
#>  $ lambda        : num [1:100] 0.289 0.276 0.263 0.251 0.24 ...
#>  $ penalty       : chr "lasso"
#>  $ family        : chr "gaussian"
#>  $ alpha         : num 1
#>  $ loss          : num [1:100] 52.9 52.3 51.7 50.6 49.2 ...
#>  $ penalty.factor: num [1:430] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ n             : int 73
#>  $ p             : num 430
#>  $ center        : num [1:430] -0.315 0.863 0.301 0 1.192 ...
#>  $ scale         : num [1:430] 5.04 4.27 5.13 4.44 4.99 ...
#>  $ y             : num [1:73] -0.268 -0.479 0.125 -0.3 0.713 ...
#>  $ col.idx       : num [1:430] 1 2 3 4 5 6 7 8 9 10 ...
#>  $ rejections    : int [1:100] 430 428 426 422 422 419 419 417 414 414 ...
#>  - attr(*, "class")= chr "big_spLinReg"

# how to use covariables?
X2 <- as.big.matrix(cbind(X[,], covar), type = "double")
test <- biglasso::biglasso(X2, y, family = "gaussian", lambda.min = 0.01,
                           alpha = 0.5, penalty = "enet")
test2 <- big_spLinReg(X, y, covar.train = covar, alpha = 0.5)
# verification
print(all.equal(test2$lambda, test$lambda))

#> [1] TRUE

print(all.equal(test2$beta@x, test$beta[-1, ]@x))

#> [1] TRUE

print(all.equal(test2$intercept, test$beta[1, ]))

#> [1] TRUE