Sparse logistic regression

Sparse logistic regression

big_spLogReg(X., y01.train, ind.train = rows_along(X.), covar.train = NULL,
  ...)

Arguments

X.
Either a big.matrix or a big.matrix.descriptor.

y01.train
Vector of responses, corresponding to ind.train. Must be only 0s and 1s.

ind.train
An optional vector of the row indices that are used, for the training part. If not specified, all rows are used. Don't use negative indices.

covar.train
Matrix of covariables to be added in each model to correct for confounders (e.g. the scores of PCA), corresponding to ind.train. Default is NULL and corresponds to only adding an intercept to each model.

...
Arguments passed on to COPY_biglasso

alpha

The elastic-net mixing parameter that controls the relative contribution from the lasso (l1) and the ridge (l2) penalty. The penalty is defined as $$ \alpha||\beta||_1 + (1-\alpha)/2||\beta||_2^2.$$ alpha = 1 is the lasso penalty and alpha in between 0 (1e-6) and 1 is the elastic-net penalty.

lambda.min

The smallest value for lambda, as a fraction of lambda.max. Default is .001 if the number of observations is larger than the number of covariates and .01 otherwise.

nlambda

The number of lambda values. Default is 100.

lambda.log.scale

Whether compute the grid values of lambda on log scale (default) or linear scale.

lambda

A user-specified sequence of lambda values. By default, a sequence of values of length nlambda is computed, equally spaced on the log scale.

eps

Convergence threshold for inner coordinate descent. The algorithm iterates until the maximum change in the objective after any coefficient update is less than eps times the null deviance. Default value is 1e-7.

max.iter

Maximum number of iterations. Default is 1000.

dfmax

Upper bound for the number of nonzero coefficients. Default is no upper bound. However, for large data sets, computational burden may be heavy for models with a large number of nonzero coefficients.

penalty.factor

A multiplicative factor for the penalty applied to each coefficient. If supplied, penalty.factor must be a numeric vector of length equal to sum of the number of columns of X and the number of covariables (intercept excluded). The purpose of penalty.factor is to apply differential penalization if some coefficients are thought to be more likely than others to be in the model. Current package doesn't allow unpenalized coefficients. That is penalty.factor cannot be 0.

warn

Return warning messages for failures to converge and model saturation? Default is TRUE.

verbose

Whether to print out the start, the timing of each lambda iteration and the end. Default is FALSE.

Value

A named list with following variables:

intercept

A vector of intercepts, corresponding to each lambda.

beta

The fitted matrix of coefficients, store in sparse matrix representation. The number of rows is equal to the number of coefficients, and the number of columns is equal to nlambda.

iter

A vector of length nlambda containing the number of iterations until convergence at each value of lambda.

lambda

The sequence of regularization parameter values in the path.

penalty

Penalty used. See the input parameter alpha.

family

Either "gaussian" or "binomial" depending on the function used.

alpha

Input parameter.

loss

A vector containing either the residual sum of squares (for linear models) or negative log-likelihood (for logistic models) of the fitted model at each value of lambda.

penalty.factor

Input parameter.

n

The number of observations used in the model fitting. It's equal to length(row.idx).

p

The number of dimensions (including covariables, but not the intercept).

center

The sample mean vector of the variables, i.e., column mean of the sub-matrix of X used for model fitting.

scale

The sample standard deviation of the variables, i.e., column standard deviation of the sub-matrix of X used for model fitting.

y

The response vector used in the model fitting. Depending on row.idx, it could be a subset of the raw input of the response vector y.

col.idx

The indices of features that have 'scale' value greater than 1e-6. Features with 'scale' less than 1e-6 are removed from model fitting.

rejections

The number of features rejected at each value of lambda.

Details

This is a modified version of one function of package biglasso. It adds the possibility to train models with covariables and use many types of big.matrix (not only double ones). Yet, it only corresponds to screen = "SSR" (Sequential Strong Rules).

References

Tibshirani, R., Bien, J., Friedman, J., Hastie, T., Simon, N., Taylor, J. and Tibshirani, R. J. (2012), Strong rules for discarding predictors in lasso-type problems. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 74: 245–266. http://dx.doi.org/10.1111/j.1467-9868.2011.01004.x.

Zeng, Y., and Breheny, P. (2016). The biglasso Package: A Memory- and Computation-Efficient Solver for Lasso Model Fitting with Big Data in R. arXiv preprint arXiv:1701.05936. https://arxiv.org/abs/1701.05936.

See also

glmnet biglasso

Examples

set.seed(1)

# simulating some data
N <- 73
M <- 430
X <- as.big.matrix(matrix(rnorm(N*M, sd = 5), N), type = "integer")
y <- sample(0:1, size = N, replace = TRUE)
covar <- matrix(rnorm(N * 3), N)

# error, only handle `double` `big.matrix` objects
not_run({
    test <- biglasso::biglasso(X, y, family = "binomial")    print(summary(test$scale))
})

# OK here
test2 <- big_spLogReg(X, y)
str(test2)

# how to use covariables?
X2 <- as.big.matrix(cbind(X[,], covar), type = "double")
test <- biglasso::biglasso(X2, y, family = "binomial", lambda.min = 0.01,
                           alpha = 0.5, penalty = "enet")
test2 <- big_spLogReg(X, y, covar.train = covar, alpha = 0.5)
# verification
print(all.equal(test2$lambda, test$lambda))
print(all.equal(test2$beta@x, test$beta[-1, ]@x))
print(all.equal(test2$intercept, test$beta[1, ]))

#> Error: <text>:12:62: unexpected symbol
#> 11: not_run({
#> 12:     test <- biglasso::biglasso(X, y, family = "binomial")    print
#>                                                                  ^