Sparse SVM

Fit solution paths for sparse linear SVM regularized by lasso or elastic-net over a grid of values for the regularization parameter lambda.

big_spSVM(X., y01.train, ind.train = rows_along(X.), covar.train = NULL,
  ...)

Arguments

X.

Either a big.matrix or a big.matrix.descriptor.

y01.train

Vector of responses, corresponding to ind.train. Must be only 0s and 1s.

ind.train

An optional vector of the row indices that are used, for the training part. If not specified, all rows are used. Don't use negative indices.

covar.train

Matrix of covariables to be added in each model to correct for confounders (e.g. the scores of PCA), corresponding to ind.train. Default is NULL and corresponds to only adding an intercept to each model.

...

Arguments passed on to COPY_sparseSVM

alpha

The elastic-net mixing parameter that controls the relative contribution from the lasso and the ridge penalty. It must be a number between 0 and 1. alpha=1 is the lasso penalty and alpha=0 the ridge penalty.

gamma

The tuning parameter for huberization smoothing of hinge loss. Default is 0.1.

nlambda

The number of lambda values. Default is 100.

lambda.min

The smallest value for lambda, as a fraction of lambda.max, the data derived entry value. Default is 0.01 if the number of observations is larger than the number of variables and 0.05 otherwise.

lambda

A user-specified sequence of lambda values. Typical usage is to leave blank and have the program automatically compute a lambda sequence based on nlambda and lambda.min. Specifying lambda overrides this. This argument should be used with care and supplied with a decreasing sequence instead of a single value.

screen

Screening rule to be applied at each lambda that discards variables for speed. Either "ASR" (default), "SR" or "none". "SR" stands for the strong rule, and "ASR" for the adaptive strong rule. Using "ASR" typically requires fewer iterations to converge than "SR", but the computing time are generally close. Note that the option "none" is used mainly for debugging, which may lead to much longer computing time.

max.iter

Maximum number of iterations. Default is 1000.

eps

Convergence threshold. The algorithms continue until the maximum change in the objective after any coefficient update is less than eps times the null deviance. Default is 1E-7.

dfmax

Upper bound for the number of nonzero coefficients. The algorithm exits and returns a partial path if dfmax is reached. Useful for very large dimensions.

penalty.factor

A numeric vector of length equal to the number of variables. Each component multiplies lambda to allow differential penalization. Can be 0 for some variables, in which case the variable is always in the model without penalization. Default is 1 for all variables.

message

If set to TRUE, sparseSVM will inform the user of its progress. Default is FALSE.

Value

A named list containing:

Details

This is a modified version of one function of https://github.com/CY-dev/sparseSVM.

See also

LiblineaR sparseSVM

Examples

# simulating some data
N <- 73
M <- 430
X <- as.big.matrix(matrix(rnorm(N*M, sd = 5), N))
y <- sample(0:1, size = N, replace = TRUE)
y.factor <- factor(y, levels = c(1, 0))
covar <- matrix(rnorm(N * 3), N)

# error, only handle standard R matrices
## Not run: test &lt;- sparseSVM::sparseSVM(X, y)

# OK here
test2 <- big_spSVM(X, y)
str(test2)

#> List of 10
#>  $ call          : language COPY_sparseSVM(X = X, y.train = y01.train, ind.train = ind.train, covar.train = covar.train)
#>  $ intercept     : Named num [1:100] -1 -0.917 -0.912 -0.901 -0.868 ...
#>   ..- attr(*, "names")= chr [1:100] "0.2732" "0.265" "0.2571" "0.2495" ...
#>  $ beta          :Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   .. ..@ i       : int [1:10594] 68 193 271 355 414 68 149 172 193 271 ...
#>   .. ..@ p       : int [1:101] 0 0 5 14 24 38 55 73 95 117 ...
#>   .. ..@ Dim     : int [1:2] 430 100
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:100] "0.2732" "0.265" "0.2571" "0.2495" ...
#>   .. ..@ x       : num [1:10594] 0.00585 -0.012478 -0.000858 -0.000889 -0.000529 ...
#>   .. ..@ factors : list()
#>  $ iter          : int [1:100] 0 4 3 4 10 57 5 9 20 5 ...
#>  $ saturated     : logi FALSE
#>  $ lambda        : num [1:100] 0.273 0.265 0.257 0.249 0.242 ...
#>  $ alpha         : num 1
#>  $ gamma         : num 0.1
#>  $ penalty.factor: num [1:430] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ levels        : int [1:2] 0 1
#>  - attr(*, "class")= chr "big_sp"

# how to use covariables?
X2 <- cbind(X[,], covar)
test <- sparseSVM::sparseSVM(X2, y.factor, alpha = 0.5)
test2 <- big_spSVM(X, y, covar.train = covar, alpha = 0.5)
# verification
print(all.equal(test2$lambda, test$lambda))

#> [1] TRUE

print(all.equal(test2$beta, test$weights[-1, ], check.attributes = FALSE))

#> [1] "Modes: S4, numeric"

print(all.equal(test2$intercept, test$weights[1, ]))

#> [1] TRUE