Tcrossprod

Compute \(X.train X.train^T\) for a big.matrix X after applying a particular scaling to it.

big_tcrossprodSelf(X, fun.scaling, ind.train = seq(nrow(X)),
  block.size = 1000, use.Eigen = !detect_MRO())

Arguments

X

A big.matrix. You shouldn't have missing values in your data.

fun.scaling

A function that returns a named list of mean and sd for every column, to scale each of their elements such as followed: $$\frac{X_{i,j} - mean_j}{sd_j}$$.

ind.train

An optional vector of the row indices that are used, for the training part. If not specified, all rows are used.

block.size

Maximum number of columns read at once. Default is 1000. This parameter controls the trade-off between memory usage and speed. Basically, the more you can load at once, the quicker will be the execution time, at the expense of memory usage.

use.Eigen

Should the Eigen library be used for matrix computations? Default tries to detect MRO. See details.

Value

A list of

\(K = X.train X.train^T\),

a numeric vector mean of column scaling,

a numeric vector sd of column scaling.

Details

For matrix computations, using Eigen library is faster. However, if you link R with an optimized math library, using R's base operations is even much faster.

For example, you can easily link R with the https://software.intel.com/en-us/intel-mkl (Intel® MKL) through https://mran.revolutionanalytics.com/open/ (MRO). It really improves performance of R and RcppArmadillo matrix computations, yet not the ones of RcppEigen (at least not directly).

So,

Eigen should be prefered if you don't change anything,

base R should be prefered if you use MRO,

Eigen may be prefered if you manage to link RcppEigen with the MKL (please mailto:florian.prive.21@gmail.com if you do!).

See also

tcrossprod

Examples

# Simulating some data
X <- big.matrix(17, 41)
X[] <- rnorm(length(X))

# Comparing with tcrossprod
big_noscale <- big_scale(center = FALSE)
test <- big_tcrossprodSelf(X, fun.scaling = big_noscale)
print(dim(test$K))

#> [1] 17 17

print(all.equal(test$K, tcrossprod(X[,])))

#> [1] TRUE

# Using only half of the data for "training"
ind <- sort(sample(nrow(X), nrow(X)/2))
test2 <- big_tcrossprodSelf(X, fun.scaling = big_noscale,
                            ind.train = ind)
print(dim(test2$K))

#> [1] 8 8

print(all.equal(test2$K, tcrossprod(X[ind, ])))

#> [1] TRUE