Transposition

This function implements a simple cache-oblivious algorithm for the transposition of a Filebacked Big Matrix.

big_transpose(X, backingfile = tempfile())

Arguments

X
A FBM.

backingfile
Path to the file storing the Big Matrix on disk. An extension ".bk" will be automatically added. Default stores in the temporary directory.

Value

The new transposed Filebacked Big Matrix (or its descriptor). Its dimensions and type are automatically determined from the input Filebacked Big Matrix.

Examples

X <- FBM(10, 5, init = rnorm(50))
X[]

#>              [,1]        [,2]        [,3]       [,4]        [,5]
#>  [1,] -0.44496684  0.04993072  1.13897648 -1.8085971  0.72536274
#>  [2,] -0.36245836 -0.32330151  0.37670599 -0.2815420  0.08113190
#>  [3,] -0.70566422  1.48164700 -0.04948612  1.6317885 -1.66166634
#>  [4,] -0.27445529 -0.08985143  1.05238058  2.2263765  0.06787402
#>  [5,] -0.58019104  0.32373714 -1.78010392  0.2249905  1.51426278
#>  [6,]  0.97313773  1.69691638 -0.20695664  1.4240952 -0.20836444
#>  [7,] -1.77959969  1.10486797  1.66767961  1.9076996 -0.20527700
#>  [8,]  0.91717523  0.83644849 -0.23942115 -0.8606823 -1.55100430
#>  [9,]  0.03515448  0.41027383 -0.16755004 -0.1033174  0.84066493
#> [10,] -0.27424994 -1.13418129  1.84122844 -0.3354759  0.18879906

Xt <- big_transpose(X)
identical(t(X[]), Xt[])

#> [1] TRUE

X <- big_attachExtdata()
Xt <- big_transpose(X)
identical(t(X[]), Xt[])

#> [1] TRUE