Column-wise linear regression

Slopes of column-wise linear regressions of each column of a big.matrix, with some other associated statistics. Covariates can be added to correct for confounders.

big_univLinReg(X., y.train, ind.train = rows_along(X.),
  ind.col = cols_along(X.), covar.train = NULL, ncores2 = 1,
  thr.eigval = 1e-04)

Arguments

X.

Either a big.matrix or a big.matrix.descriptor.

y.train

Vector of responses, corresponding to ind.train.

ind.train

An optional vector of the row indices that are used, for the training part. If not specified, all rows are used. Don't use negative indices.

ind.col

An optional vector of the column indices that are used. If not specified, all columns are used. Don't use negative indices.

covar.train

Matrix of covariables to be added in each model to correct for confounders (e.g. the scores of PCA), corresponding to ind.train. Default is NULL and corresponds to only adding an intercept to each model.

ncores2

Number of cores used. Default doesn't use parallelism. For this function, use only half of the cores you have.

thr.eigval

Threshold to remove "unsignificant" singular vectors. Default is 1e-4.

Value

A data.frame with 4 elements:

the slopes of each regression,

the standard errors of each slope,

the t-scores associated with each slope,

the p-values associated with each t-score.

See also

lm

Examples

X.desc <- big_attachExtdata()
n <- nrow(X.desc)
y <- rnorm(n)
covar <- matrix(rnorm(n * 3), n)

X1 <- attach.BM(X.desc)[, 1] # only first column of the `big.matrix`

# Without covar
str(big_univLinReg(X.desc, y))

#> 'data.frame':	4542 obs. of  4 variables:
#>  $ estim  : num  0.05998 0.02767 0.13799 0.04534 -0.00702 ...
#>  $ std.err: num  0.067 0.08 0.0724 0.0816 0.0633 ...
#>  $ t.score: num  0.895 0.346 1.907 0.556 -0.111 ...
#>  $ p.value: num  0.371 0.729 0.057 0.579 0.912 ...

summary(lm(y ~ X1))$coefficients[2, ]

#>   Estimate Std. Error    t value   Pr(>|t|) 
#> 0.05997866 0.06704274 0.89463320 0.37140111 

# With all data
str(big_univLinReg(X.desc, y, covar = covar))

#> 'data.frame':	4542 obs. of  4 variables:
#>  $ estim  : num  0.05095 0.01398 0.14294 0.03454 -0.00999 ...
#>  $ std.err: num  0.0673 0.0804 0.0725 0.0816 0.0632 ...
#>  $ t.score: num  0.757 0.174 1.972 0.423 -0.158 ...
#>  $ p.value: num  0.4496 0.8619 0.0491 0.6723 0.8745 ...

summary(lm(y ~ X1 + covar))$coefficients[2, ]

#>   Estimate Std. Error    t value   Pr(>|t|) 
#> 0.05095444 0.06734813 0.75658287 0.44964780 

# With only half of the data
ind.train <- sort(sample(n, n/2))
str(big_univLinReg(X.desc, y[ind.train],
                   covar.train = covar[ind.train, ],
                   ind.train = ind.train))

#> 'data.frame':	4542 obs. of  4 variables:
#>  $ estim  : num  0.1534 -0.1626 0.0871 0.0847 0.0131 ...
#>  $ std.err: num  0.096 0.1123 0.1103 0.1146 0.0918 ...
#>  $ t.score: num  1.599 -1.448 0.79 0.739 0.143 ...
#>  $ p.value: num  0.111 0.149 0.43 0.461 0.887 ...

summary(lm(y ~ X1 + covar, subset = ind.train))$coefficients[2, ]

#>   Estimate Std. Error    t value   Pr(>|t|) 
#> 0.15344913 0.09596012 1.59909284 0.11104802