Column-wise logistic regression

Slopes of column-wise logistic regressions of each column of a big.matrix, with some other associated statistics. Covariates can be added to correct for confounders.

big_univLogReg(X, y01.train, ind.train = seq(nrow(X)), covar.train = NULL,
  ncores2 = 1, tol = 1e-08, maxiter = 20)

Arguments

X

A big.matrix. You shouldn't have missing values in your data.

y01.train

Vector of responses, corresponding to ind.train. Must be 0s and 1s.

ind.train

An optional vector of the row indices that are used, for the training part. If not specified, all rows are used.

covar.train

Matrix of covariables to be added in each model to correct for confounders (e.g. the scores of PCA), corresponding to ind.train. Default is NULL and corresponds to only adding an intercept to each model.

ncores2

Number of cores used. Default doesn't use parallelism. For this function, use only half of the cores you have.

tol

Relative tolerance to assess convergence of the coefficient. Default is 1e-8.

maxiter

Maximum number of iterations before giving up. Default is 20. Usually, convergence is reached within 3 or 4 iterations. If there is not convergence, glm is used instead for the corresponding column.

Value

A data.frame with as elements:

the slopes of each regression,

the standard errors of each slope,

the number of iteration for each slope. If is NA, this means that the algorithm didn't converge, and glm was used instead.

the z-scores associated with each slope,

the p-values associated with each z-score.

See also

glm

Examples

# Simulating some data
data("trees")
N <- nrow(trees)
covar <- matrix(rnorm(N * 3), N)
X <- as.big.matrix(as.matrix(trees))
y <- sample(0:1, size = N, replace = TRUE)

# Without covar
print(big_univLogReg(X, y))

#>         estim    std.err niter    z.score   p.value
#> 1 -0.13530714 0.12196104     4 -1.1094292 0.2672451
#> 2 -0.02037495 0.05857651     4 -0.3478348 0.7279643
#> 3 -0.02311041 0.02311723     4 -0.9997053 0.3174532

for (i in 1:3) {
  print(summary(glm(y ~ trees[, i], family = "binomial"))$coefficients[2, ])
}

#>   Estimate Std. Error    z value   Pr(>|z|) 
#> -0.1353071  0.1219610 -1.1094294  0.2672450 
#>    Estimate  Std. Error     z value    Pr(>|z|) 
#> -0.02037495  0.05857651 -0.34783477  0.72796428 
#>    Estimate  Std. Error     z value    Pr(>|z|) 
#> -0.02311041  0.02311722 -0.99970536  0.31745312 

# With all data
print(big_univLogReg(X, y, covar.train = covar))

#>         estim    std.err niter    z.score   p.value
#> 1 -0.11584156 0.12790470     4 -0.9056865 0.3651018
#> 2 -0.01119162 0.06174302     4 -0.1812613 0.8561625
#> 3 -0.01826804 0.02461764     4 -0.7420710 0.4580443

for (i in 1:3) {
  print(summary(glm(y ~ trees[, i] + covar,
                    family = "binomial"))$coefficients[2, ])
}

#>   Estimate Std. Error    z value   Pr(>|z|) 
#> -0.1158416  0.1279047 -0.9056867  0.3651017 
#>    Estimate  Std. Error     z value    Pr(>|z|) 
#> -0.01119162  0.06174300 -0.18126136  0.85616243 
#>    Estimate  Std. Error     z value    Pr(>|z|) 
#> -0.01826804  0.02461764 -0.74207108  0.45804427 

# With only half of the data
ind.train <- sort(sample(N, N / 2))

print(big_univLogReg(X, y[ind.train], covar.train = covar[ind.train, ],
                     ind.train = ind.train))

#>         estim    std.err niter    z.score   p.value
#> 1 -0.15522201 0.20501404     5 -0.7571287 0.4489728
#> 2 -0.68590808 0.55056883     9 -1.2458171 0.2128315
#> 3 -0.03743351 0.03754174     5 -0.9971169 0.3187078

for (i in 1:3) {
  print(summary(glm(y ~ trees[, i] + covar, subset = ind.train,
                    family = "binomial"))$coefficients[2, ])
}

#>   Estimate Std. Error    z value   Pr(>|z|) 
#> -0.1552220  0.2050044 -0.7571642  0.4489515 
#>   Estimate Std. Error    z value   Pr(>|z|) 
#> -0.6859081  0.5505452 -1.2458707  0.2128119 
#>    Estimate  Std. Error     z value    Pr(>|z|) 
#> -0.03743351  0.03754174 -0.99711698  0.31870773