Univariate logistic regression

Slopes of univariate logistic regressions of each column of a big.matrix, with some other associated statistics. Covariates can be added to correct for confounders.

big_univRegLog(X, y01.train, ind.train = seq(nrow(X)), covar.train = NULL,
  ncores = 1, tol = 1e-08, maxiter = 20)

Arguments

X

A big.matrix. You shouldn't have missing values in your data.

y01.train

Vector of responses, corresponding to ind.train. Must be 0s and 1s.

ind.train

An optional vector of the row indices that are used, for the training part. If not specified, all rows are used.

covar.train

Matrix of covariables to be added in each model to correct for confounders (e.g. the scores of PCA), corresponding to ind.train. Default is NULL and corresponds to only adding an Intercept to each model.

ncores

Number or cores used. Default doesn't use parallelism.

tol

Relative tolerance to assess convergence of the coefficient. Default is 1e-8.

maxiter

Maximum number of iterations before giving up. Default is 20. Usually, convergence is reached within 3 or 4 iterations. If there is not convergence, glm is used instead for the corresponding column.

Value

A data.frame with as elements:

the slopes of each regression,

the standard errors of each slope,

the number of iteration for each slope. If is NA, this means that the algorithm didn't converge, and glm was used instead.

the z-scores associated with each slope,

the p-values associated with each z-score.

See also

glm

Examples

# Simulating some data
data("trees")
N <- nrow(trees)
covar <- matrix(rnorm(N * 3), N)
X <- as.big.matrix(as.matrix(trees))
y <- sample(0:1, size = N, replace = TRUE)

# without covar
print(big_univRegLog(X, y))

#>         estim    std.err niter    z.score   p.value
#> 1 -0.05512883 0.12750440     5 -0.4323681 0.6654739
#> 2  0.01514962 0.06184304     4  0.2449689 0.8064805
#> 3 -0.01760860 0.02580653     5 -0.6823312 0.4950295

for (i in 1:3) {
  print(summary(glm(y ~ trees[, i], family = "binomial"))$coefficients[2, ])
}

#>    Estimate  Std. Error     z value    Pr(>|z|) 
#> -0.05512883  0.12750433 -0.43236829  0.66547376 
#>   Estimate Std. Error    z value   Pr(>|z|) 
#> 0.01514962 0.06184302 0.24496903 0.80648041 
#>    Estimate  Std. Error     z value    Pr(>|z|) 
#> -0.01760860  0.02580651 -0.68233159  0.49502930 

# With all data
print(big_univRegLog(X, y, covar.train = covar))

#>          estim    std.err niter    z.score   p.value
#> 1  0.017504134 0.14243452     4  0.1228925 0.9021922
#> 2  0.073971150 0.07445924     5  0.9934449 0.3204932
#> 3 -0.003229721 0.02748459     4 -0.1175102 0.9064557

for (i in 1:3) {
  print(summary(glm(y ~ trees[, i] + covar,
                    family = "binomial"))$coefficients[2, ])
}

#>   Estimate Std. Error    z value   Pr(>|z|) 
#> 0.01750413 0.14243450 0.12289251 0.90219222 
#>   Estimate Std. Error    z value   Pr(>|z|) 
#> 0.07397115 0.07445910 0.99344671 0.32049231 
#>     Estimate   Std. Error      z value     Pr(>|z|) 
#> -0.003229721  0.027484585 -0.117510265  0.906455710 

# With only half of the data
ind.train <- sort(sample(N, N / 2))

print(big_univRegLog(X, y[ind.train], covar.train = covar[ind.train, ],
                     ind.train = ind.train))

#>       estim   std.err niter  z.score   p.value
#> 1 0.6015489 0.5236546     7 1.148751 0.2506585
#> 2 0.5171095 0.4428201     9 1.167764 0.2429018
#> 3 0.1451546 0.1336804     8 1.085833 0.2775529

for (i in 1:3) {
  print(summary(glm(y ~ trees[, i] + covar, subset = ind.train,
                    family = "binomial"))$coefficients[2, ])
}

#>   Estimate Std. Error    z value   Pr(>|z|) 
#>  0.6015489  0.5236541  1.1487525  0.2506581 
#>   Estimate Std. Error    z value   Pr(>|z|) 
#>  0.5171095  0.4428197  1.1677654  0.2429014 
#>   Estimate Std. Error    z value   Pr(>|z|) 
#>  0.1451546  0.1336751  1.0858759  0.2775339