bigstatsr: Statistical Tools for Bigmemory Matrices

Easy-to-use, efficient, flexible and scalable statistical tools. Use bigmemory matrices.
Arguments

X 	A big.matrix.

X.desc 	A big.matrix.descriptor.

X. 	Either a big.matrix or a big.matrix.descriptor.

X.code 	Either a BM.code or a BM.code.descriptor.

y.train 	Vector of responses, corresponding to ind.train.

y01.train 	Vector of responses, corresponding to ind.train. Must be only 0s and 1s.

ind.train 	An optional vector of the row indices that are used, for the training part. If not specified, all rows are used. Don't use negative indices.

ind.row 	An optional vector of the row indices that are used. If not specified, all rows are used. Don't use negative indices.

ind.col 	An optional vector of the column indices that are used. If not specified, all columns are used. Don't use negative indices.

block.size 	Maximum number of columns read at once. Default is 1000.

ncores 	Number of cores used. Default doesn't use parallelism.

ncores2 	Number of cores used. Default doesn't use parallelism. For this function, use only half of the cores you have.

fun.scaling 	A function that returns a named list of mean and sd for every column, to scale each of their elements such as followed: $$\frac{X_{i,j} - mean_j}{sd_j}$$.

covar.train 	Matrix of covariables to be added in each model to correct for confounders (e.g. the scores of PCA), corresponding to ind.train. Default is NULL and corresponds to only adding an intercept to each model.

covar.row 	Matrix of covariables to be added in each model to correct for confounders (e.g. the scores of PCA), corresponding to ind.row. Default is NULL and corresponds to only adding an intercept to each model.

thr.eigval 	Threshold to remove "unsignificant" singular vectors. Default is 1e-4.

See also
Useful links:

https://privefl.github.io/bigstatsr

Report bugs at https://github.com/privefl/bigstatsr/issues