Standard statistics for

Standard statistics for columns of a big.matrix. For now, the sum, mean, sd and var are implemented.

colsums(X, ind.train = seq_len(nrow(X)))

colmeans(X, ind.train = seq_len(nrow(X)))

colvars(X, ind.train = seq_len(nrow(X)))

colsds(X, ind.train = seq_len(nrow(X)))

colmeans_sds(X, ind.train = seq_len(nrow(X)))

Arguments

X

A big.matrix. You shouldn't have missing values in your data.

ind.train

An optional vector of the row indices that are used, for the training part. If not specified, all data are used.

Value

Numeric vectors with the column statistics.

Examples

# Simulating some data
X <- big.matrix(10, 10, shared = FALSE)
X[] <- rnorm(100)

# Check the results
print(test <- colsums(X))

#>  [1] -3.6583651  0.1382521  1.6919585 -1.4090388  2.4115477 -2.0479979
#>  [7] -2.1526717 -2.0635185 -3.7271186 -0.1131829

print(test - colSums(X[,]))

#>  [1] -4.440892e-16 -2.775558e-17  0.000000e+00 -2.220446e-16 -8.881784e-16
#>  [6]  0.000000e+00  0.000000e+00  0.000000e+00  4.440892e-16 -1.110223e-16

# Only with the first 5 rows
ind <- 1:5
print(test <- colsums(X, ind))

#>  [1] -1.2739234  0.5478922 -0.1445164 -1.6588057  2.2547740 -1.9298952
#>  [7] -0.4208974  1.7170870 -3.8654257 -1.8007172

print(test - colSums(X[ind, ]))

#>  [1]  0.000000e+00  0.000000e+00  8.326673e-17 -2.220446e-16  0.000000e+00
#>  [6] -2.220446e-16  0.000000e+00  0.000000e+00  4.440892e-16  0.000000e+00

# NA in, NA out
X[2, 5] <- NA
print(colsums(X))

#>  [1] -3.6583651  0.1382521  1.6919585 -1.4090388         NA -2.0479979
#>  [7] -2.1526717 -2.0635185 -3.7271186 -0.1131829