Functions for creating a "big.matrix"

The 3 following functions return functions as well:

BM: wrapper around big.matrix.

FBM: wrapper around filebacked.big.matrix. In particular:

backingpath is "backingfiles" by default,

extension .bk is used,

descriptor is never binary and has an extension ".desc",

the big.matrix has no dimname,

option separated is never used.

tmpFBM: wrapper around FBM to get a temporary filebacked big.matrix. It uses tempfile to get the location of the temporary filebacked big.matrix. Temporary means not remaining if you restart your R session.

BM(..., descriptor = TRUE)

FBM(backingroot, backingpath = "backingfiles", init = NULL,
  descriptor = TRUE)

tmpFBM(init = NULL, descriptor = TRUE)

Arguments

...
Arguments passed on to bigmemory::big.matrix

init

a scalar value for initializing the matrix (NULL by default to avoid unnecessary time spent doing the initializing).

dimnames

a list of the row and column names; use with caution for large objects.

separated

use separated column organization of the data; see details.

backingfile

the root name for the file(s) for the cache of x.

backingpath

the path to the directory containing the file backing cache.

descriptorfile

the name of the file to hold the backingfile description, for subsequent use with attach.big.matrix; if NULL, the backingfile is used as the root part of the descriptor file name. The descriptor file is placed in the same directory as the backing files.

binarydescriptor

the flag to specify if the binary RDS format should be used for the backingfile description, for subsequent use with attach.big.matrix; if NULL of FALSE, the dput() file format is used.

shared

TRUE by default, and always TRUE if the big.matrix is file-backed. For a non-filebacked big.matrix, shared=FALSE uses non-shared memory, which can be more stable for large (say, >50% of RAM) objects. Shared memory allocation can sometimes fail in such cases due to exhausted shared-memory resources in the system.

descriptor
Return the descriptor of the resulting big.matrix or directly the big.matrix object? Default is TRUE (the descriptor).

backingroot
The root of the backing file to be created. This should not be a full path (use backingpath) nor have any extension (".bk" is used by default).

backingpath
The directory of the backing file to be created. Default is "backingfiles". The directory is created if doesn't exist.

init
A scalar value for initializing the matrix (NULL by default to avoid unnecessary time spent doing the initializing).

Value

A function of 3 parameters nrow, ncol and type (default is "double") that is used to create a big.matrix. This is particularly useful as the fun.createBM parameter.

Examples

# these functions return **functions**
(f <- BM(shared = FALSE, descriptor = FALSE))

#> function(nrow, ncol, type = "double") {
#>     res <- big.matrix(nrow, ncol, type, ...)
#>     `if`(descriptor, describe(res), res)
#>   }
#> <environment: 0xccf3280>

# with 3 parameters
(X <- f(nrow = 10, ncol = 6, type = "integer"))

#> An object of class "big.matrix"
#> Slot "address":
#> <pointer: 0xdddac90>
#> 

is.shared(X)

#> [1] FALSE

# a shortcut for a filebacked `big.matrix`
X2.desc <- FBM(backingroot = "test", init = 2)(10, 10)
str(X2.desc)

#> Formal class 'big.matrix.descriptor' [package "bigmemory"] with 1 slot
#>   ..@ description:List of 13
#>   .. ..$ sharedType: chr "FileBacked"
#>   .. ..$ filename  : chr "test.bk"
#>   .. ..$ dirname   : chr "backingfiles/"
#>   .. ..$ totalRows : num 10
#>   .. ..$ totalCols : num 10
#>   .. ..$ rowOffset : num 0
#>   .. ..$ colOffset : num 0
#>   .. ..$ nrow      : num 10
#>   .. ..$ ncol      : num 10
#>   .. ..$ rowNames  : NULL
#>   .. ..$ colNames  : NULL
#>   .. ..$ type      : chr "double"
#>   .. ..$ separated : logi FALSE

attach.BM(X2.desc)[,]

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    2    2    2    2    2    2    2    2    2     2
#>  [2,]    2    2    2    2    2    2    2    2    2     2
#>  [3,]    2    2    2    2    2    2    2    2    2     2
#>  [4,]    2    2    2    2    2    2    2    2    2     2
#>  [5,]    2    2    2    2    2    2    2    2    2     2
#>  [6,]    2    2    2    2    2    2    2    2    2     2
#>  [7,]    2    2    2    2    2    2    2    2    2     2
#>  [8,]    2    2    2    2    2    2    2    2    2     2
#>  [9,]    2    2    2    2    2    2    2    2    2     2
#> [10,]    2    2    2    2    2    2    2    2    2     2

unlink(paste0("backingfiles/test", c(".bk", ".desc")))

# a temporary filebacked `big.matrix`
test <- tmpFBM()(10, 10)
str(test)

#> Formal class 'big.matrix.descriptor' [package "bigmemory"] with 1 slot
#>   ..@ description:List of 13
#>   .. ..$ sharedType: chr "FileBacked"
#>   .. ..$ filename  : chr "file4d512893c6a6.bk"
#>   .. ..$ dirname   : chr "/tmp/RtmpTLT3SZ/"
#>   .. ..$ totalRows : num 10
#>   .. ..$ totalCols : num 10
#>   .. ..$ rowOffset : num 0
#>   .. ..$ colOffset : num 0
#>   .. ..$ nrow      : num 10
#>   .. ..$ ncol      : num 10
#>   .. ..$ rowNames  : NULL
#>   .. ..$ colNames  : NULL
#>   .. ..$ type      : chr "double"
#>   .. ..$ separated : logi FALSE