Predict method

Predict method for class big_sp.
# S3 method for big_sp
predict(object, X., ind.row = rows_along(X.),
  covar.row = NULL, block.size = 1000, ...)

Arguments

object 	Object of class big_sp.

X. 	Either a big.matrix or a big.matrix.descriptor.

ind.row 	An optional vector of the row indices that are used. If not specified, all rows are used. Don't use negative indices.

covar.row 	Matrix of covariables to be added in each model to correct for confounders (e.g. the scores of PCA), corresponding to ind.row. Default is NULL and corresponds to only adding an intercept to each model.

block.size 	Maximum number of columns read at once. Default is 1000.

... 	Not used.

Value
A matrix of scores, with rows corresponding to ind.row and columns corresponding to lambda.
See also
big_spLinReg, big_spLogReg and big_spSVM.
Examples

# simulating some data
N <- 73
M <- 430
X.desc <- describe(as.big.matrix(matrix(rnorm(N*M, sd = 5), N)))
y <- sample(0:1, size = N, replace = TRUE)
covar <- matrix(rnorm(N * 3), N)

ind.train <- sort(sample(N, N/2))
mod <- big_spSVM(X.desc, y[ind.train], ind.train = ind.train,
                  covar.train = covar[ind.train, ])
ind.test <- setdiff(rows_along(X.desc), ind.train)
pred <- predict(mod, X. = X.desc, ind.row = ind.test,
                covar.row = covar[ind.test, ])