Predict method

Predict method for class big_sp_list.

# S3 method for big_sp_list
predict(
  object,
  X,
  ind.row = rows_along(X),
  ind.col = attr(object, "ind.col"),
  covar.row = NULL,
  proba = (attr(object, "family") == "binomial"),
  base.row = NULL,
  ncores = 1,
  ...
)

Arguments

object

Object of class big_sp_list.

X

An object of class FBM.

ind.row

An optional vector of the row indices that are used. If not specified, all rows are used. Don't use negative indices.

ind.col

An optional vector of the column indices that are used. If not specified, all columns are used. Don't use negative indices.

covar.row

Matrix of covariables to be added in each model to correct for confounders (e.g. the scores of PCA), corresponding to ind.row. Default is NULL and corresponds to only adding an intercept to each model. You can use covar_from_df() to convert from a data frame.

proba

Whether to return probabilities?

base.row

Vector of base predictions, corresponding to ind.row.

ncores

Number of cores used. Default doesn't use parallelism. You may use nb_cores.

...

Not used.

Value

A vector of scores, corresponding to ind.row.

See also