Remove temporary backing files

Remove temporary backing files

tmpFBM.rm(X.)

Arguments

X.
Either a big.matrix or a big.matrix.descriptor.

Value

These functions return a logical vector indicating which operation succeeded for each of the files attempted. Using a missing value for a file or path name will always be regarded as a failure.

If showWarnings = TRUE, file.create will give a warning for an unexpected failure.

Examples

list.files(path = tempdir())

#>  [1] "269228147a2947dd909c079dc4ea8b35.png"                       
#>  [2] "5bd9c8bf705e4daeb2593fc0f8b53496.png"                       
#>  [3] "6f34ff9c74014ecf9c31f295c8fac688.png"                       
#>  [4] "72ddf8411022405c92a447241761e9b4.png"                       
#>  [5] "7ecc0b32fa8548929ba16b717a57789c.png"                       
#>  [6] "80c9bee99ad649e49e8166b241717d24.png"                       
#>  [7] "b16d535e94634daaad866e41db258ee1.png"                       
#>  [8] "d5623a7f6a5b498f85f7bef501fb3150.png"                       
#>  [9] "e6c03f9dfd4448abb6d4084bda3a1166.png"                       
#> [10] "ed4756ae59c94b12a47dd55e43e51807.png"                       
#> [11] "file268a215e33d3.bk"                                        
#> [12] "file268a215e33d3.desc"                                      
#> [13] "file268a524c3761.bk"                                        
#> [14] "file268a524c3761.desc"                                      
#> [15] "file268a7c1dab5f.bk"                                        
#> [16] "file268a7c1dab5f.desc"                                      
#> [17] "gitmsg268a2715e864.txt"                                     
#> [18] "repos_https%3A%2F%2Fcran.rstudio.com%2F%2Fsrc%2Fcontrib.rds"

X.desc <- tmpFBM()(10, 10)
list.files(path = tempdir())

#>  [1] "269228147a2947dd909c079dc4ea8b35.png"                       
#>  [2] "5bd9c8bf705e4daeb2593fc0f8b53496.png"                       
#>  [3] "6f34ff9c74014ecf9c31f295c8fac688.png"                       
#>  [4] "72ddf8411022405c92a447241761e9b4.png"                       
#>  [5] "7ecc0b32fa8548929ba16b717a57789c.png"                       
#>  [6] "80c9bee99ad649e49e8166b241717d24.png"                       
#>  [7] "b16d535e94634daaad866e41db258ee1.png"                       
#>  [8] "d5623a7f6a5b498f85f7bef501fb3150.png"                       
#>  [9] "e6c03f9dfd4448abb6d4084bda3a1166.png"                       
#> [10] "ed4756ae59c94b12a47dd55e43e51807.png"                       
#> [11] "file268a215e33d3.bk"                                        
#> [12] "file268a215e33d3.desc"                                      
#> [13] "file268a524c3761.bk"                                        
#> [14] "file268a524c3761.desc"                                      
#> [15] "file268a694cfb22.bk"                                        
#> [16] "file268a694cfb22.desc"                                      
#> [17] "file268a7c1dab5f.bk"                                        
#> [18] "file268a7c1dab5f.desc"                                      
#> [19] "gitmsg268a2715e864.txt"                                     
#> [20] "repos_https%3A%2F%2Fcran.rstudio.com%2F%2Fsrc%2Fcontrib.rds"

tmpFBM.rm(X.desc)

#> [1] TRUE TRUE

list.files(path = tempdir())

#>  [1] "269228147a2947dd909c079dc4ea8b35.png"                       
#>  [2] "5bd9c8bf705e4daeb2593fc0f8b53496.png"                       
#>  [3] "6f34ff9c74014ecf9c31f295c8fac688.png"                       
#>  [4] "72ddf8411022405c92a447241761e9b4.png"                       
#>  [5] "7ecc0b32fa8548929ba16b717a57789c.png"                       
#>  [6] "80c9bee99ad649e49e8166b241717d24.png"                       
#>  [7] "b16d535e94634daaad866e41db258ee1.png"                       
#>  [8] "d5623a7f6a5b498f85f7bef501fb3150.png"                       
#>  [9] "e6c03f9dfd4448abb6d4084bda3a1166.png"                       
#> [10] "ed4756ae59c94b12a47dd55e43e51807.png"                       
#> [11] "file268a215e33d3.bk"                                        
#> [12] "file268a215e33d3.desc"                                      
#> [13] "file268a524c3761.bk"                                        
#> [14] "file268a524c3761.desc"                                      
#> [15] "file268a7c1dab5f.bk"                                        
#> [16] "file268a7c1dab5f.desc"                                      
#> [17] "gitmsg268a2715e864.txt"                                     
#> [18] "repos_https%3A%2F%2Fcran.rstudio.com%2F%2Fsrc%2Fcontrib.rds"
#> [19] "rs-graphics-170e53dc-5ebd-4bd1-9e32-94525ef9b2f2"