AUC

Compute the Area Under the ROC Curve (AUC) of a predictor and possibly its 95% confidence interval.

AUC(pred, target, digits = NULL)

AUCBoot(pred, target, nboot = 10000, seed = NA, digits = NULL)

Arguments

pred

Vector of predictions.

target

Vector of true labels (must have exactly two levels, no missing values).

digits

See round. Default doesn't use rounding.

nboot

Number of bootstrap samples to evaluate the 95% CI. Default is 1e4.

seed

See set.seed. Use it for reproducibility. Default doesn't set any seed.

Value

The AUC, a probability, and possibly its 2.5% and 97.5% quantiles (95% CI).

Details

Other packages provide ways to compute the AUC (see this http://stats.stackexchange.com/a/146174/135793). I chose to compute the AUC through its statistical definition as a probability: $$P(score(x_{case}) > score(x_{control})).$$ Note that I consider equality between scores as a 50%-probability of one being greater than the other.

References

Hanley, J. A., & McNeil, B. J. (1982). The meaning and use of the area under a receiver operating characteristic (ROC) curve. Radiology, 143(1), 29-36. http://dx.doi.org/10.1148/radiology.143.1.7063747.

Tom Fawcett. 2006. An introduction to ROC analysis. Pattern Recogn. Lett. 27, 8 (June 2006), 861-874. http://dx.doi.org/10.1016/j.patrec.2005.10.010.

See also

wilcox.test

Examples

set.seed(1)

AUC(c(0, 0), 0:1) # Equality of scores

#> [1] 0.5

AUC(c(0.2, 0.1, 1), c(-1, -1, 1)) # Perfect AUC

#> [1] 1

x <- rnorm(100)
z <- rnorm(length(x), x, abs(x))
y <- sign(z)
print(AUC(x, y))

#> [1] 0.8071659

print(AUCBoot(x, y))

#>       Mean       2.5%      97.5%         Sd 
#> 0.80706332 0.70504945 0.89642714 0.04852244