Sparse regression path

Fit solution paths for linear or logistic regression models penalized by lasso (alpha = 1) or elastic-net (1e-6 < alpha < 1) over a grid of values for the regularization parameter lambda.

COPY_biglasso(X, y.train, ind.train, covar.train, family = c("gaussian",
  "binomial"), alpha = 1, lambda.min = if (nrow(X) > ncol(X)) 0.001 else
  0.01, nlambda = 100, lambda.log.scale = TRUE, lambda, eps = 1e-07,
  max.iter = 1000, dfmax = p + 1, penalty.factor = NULL, warn = TRUE,
  verbose = FALSE)

Arguments

family
Either "gaussian" or "binomial", depending on the response.

alpha
The elastic-net mixing parameter that controls the relative contribution from the lasso (l1) and the ridge (l2) penalty. The penalty is defined as $$ \alpha||\beta||_1 + (1-\alpha)/2||\beta||_2^2.$$ alpha = 1 is the lasso penalty and alpha in between 0 (1e-6) and 1 is the elastic-net penalty.

lambda.min
The smallest value for lambda, as a fraction of lambda.max. Default is .001 if the number of observations is larger than the number of covariates and .01 otherwise.

nlambda
The number of lambda values. Default is 100.

lambda.log.scale
Whether compute the grid values of lambda on log scale (default) or linear scale.

lambda
A user-specified sequence of lambda values. By default, a sequence of values of length nlambda is computed, equally spaced on the log scale.

eps
Convergence threshold for inner coordinate descent. The algorithm iterates until the maximum change in the objective after any coefficient update is less than eps times the null deviance. Default value is 1e-7.

max.iter
Maximum number of iterations. Default is 1000.

dfmax
Upper bound for the number of nonzero coefficients. Default is no upper bound. However, for large data sets, computational burden may be heavy for models with a large number of nonzero coefficients.

penalty.factor
A multiplicative factor for the penalty applied to each coefficient. If supplied, penalty.factor must be a numeric vector of length equal to sum of the number of columns of X and the number of covariables (intercept excluded). The purpose of penalty.factor is to apply differential penalization if some coefficients are thought to be more likely than others to be in the model. Current package doesn't allow unpenalized coefficients. That is penalty.factor cannot be 0.

warn
Return warning messages for failures to converge and model saturation? Default is TRUE.

verbose
Whether to print out the start, the timing of each lambda iteration and the end. Default is FALSE.

Value

A named list with following variables:

intercept

A vector of intercepts, corresponding to each lambda.

beta

The fitted matrix of coefficients, store in sparse matrix representation. The number of rows is equal to the number of coefficients, and the number of columns is equal to nlambda.

iter

A vector of length nlambda containing the number of iterations until convergence at each value of lambda.

lambda

The sequence of regularization parameter values in the path.

penalty

Penalty used. See the input parameter alpha.

family

Either "gaussian" or "binomial" depending on the function used.

alpha

Input parameter.

loss

A vector containing either the residual sum of squares (for linear models) or negative log-likelihood (for logistic models) of the fitted model at each value of lambda.

penalty.factor

Input parameter.

n

The number of observations used in the model fitting. It's equal to length(row.idx).

p

The number of dimensions (including covariables, but not the intercept).

center

The sample mean vector of the variables, i.e., column mean of the sub-matrix of X used for model fitting.

scale

The sample standard deviation of the variables, i.e., column standard deviation of the sub-matrix of X used for model fitting.

y

The response vector used in the model fitting. Depending on row.idx, it could be a subset of the raw input of the response vector y.

col.idx

The indices of features that have 'scale' value greater than 1e-6. Features with 'scale' less than 1e-6 are removed from model fitting.

rejections

The number of features rejected at each value of lambda.

Details

The objective function for linear regression (family = "gaussian") is $$\frac{1}{2n}\textrm{RSS} + \textrm{penalty},$$ for logistic regression (family = "binomial") it is $$-\frac{1}{n} loglike + \textrm{penalty}.$$