Train one model

Train one model

COPY_biglasso_part(X, y.train, ind.train, ind.col, covar.train, family, lambda,
  center, scale, resid, alpha, eps, max.iter, dfmax, warn, ind.val, covar.val,
  y.val, n.abort, nlam.min)

Value

A named list with following variables:

intercept

A vector of intercepts, corresponding to each lambda.

beta

The fitted matrix of coefficients, store in sparse matrix representation. The number of rows is equal to the number of coefficients, and the number of columns is equal to nlambda.

iter

A vector of length nlambda containing the number of iterations until convergence at each value of lambda.

lambda

The sequence of regularization parameter values in the path.

family

Either "gaussian" or "binomial" depending on the function used.

alpha

Input parameter.

loss

A vector containing either the residual sum of squares (for linear models) or negative log-likelihood (for logistic models) of the fitted model at each value of lambda.

loss.val

A vector containing the loss for the corresponding validation set.

n

The number of observations used in the model fitting. It's equal to length(row.idx).

p

The number of dimensions (including covariables, but not the intercept).

center

The sample mean vector of the variables, i.e., column mean of the sub-matrix of X used for model fitting.

scale

The sample standard deviation of the variables, i.e., column standard deviation of the sub-matrix of X used for model fitting.

y

The response vector used in the model fitting. Depending on row.idx, it could be a subset of the raw input of the response vector y.

col.idx

The indices of features that have 'scale' value greater than 1e-6. Features with 'scale' less than 1e-6 are removed from model fitting.