Class FBM

A reference class for storing and accessing matrix-like data stored in files on disk. This is very similar to Filebacked Big Matrices provided by the bigmemory package. Yet, the implementation is much more lighter.

Wrapper constructor for class FBM.

FBM(nrow, ncol, type = c("double", "integer", "unsigned short",
  "unsigned char", "raw"), init = NULL, backingfile = tempfile(),
  create_bk = TRUE, save = FALSE)

Arguments

nrow
Number of rows.

ncol
Number of columns.

type
Type of the Filebacked Big Matrix (default is double). Either

"double"

"integer"

"unsigned short": can store integer values from 0 to 65535. It has vocation to become the basis for a FBM.code65536 class for accessing strings.

"raw" or "unsigned char": can store integer values from 0 to 255. It is the basis for the FBM.code256 class for accessing 256 arbitrary different numeric values. It is used in package bigsnpr.

init
Either a single value (e.g. 0) or as many value as the number of elements of the FBM. Default doesn't initialize the matrix.

backingfile
Path to the file storing the Big Matrix on disk. An extension ".bk" will be automatically added. Default stores in the temporary directory.

create_bk
Create a backingfile (the default) or use an existing one (which should be named by the backingfile parameter and have an extension ".bk"). For example, this could be used to convert a filebacked big.matrix from package bigmemory to a FBM.

save
Whether to save the result object in an ".rds" file alongside the backingfile. Default is FALSE.

Examples

X <- FBM(10, 10)
typeof(X)

#> [1] "double"

X[] <- rnorm(length(X))
X[, 1:6]

#>               [,1]       [,2]       [,3]        [,4]        [,5]         [,6]
#>  [1,] -1.791337818 -0.8432650  1.1651380  1.13199575  1.35224643  0.929763401
#>  [2,] -3.045363930 -0.6840283 -0.2783777 -0.02443501  0.78862870 -1.290008417
#>  [3,] -0.097388392  0.8459935 -0.5468960 -0.89638534 -2.24788016 -0.570331658
#>  [4,] -1.296783557 -0.6240430 -0.2793260 -0.39005692  0.04787734  0.961281650
#>  [5,]  0.191180067  0.8847433  0.2120864 -1.07864812  0.88008117 -0.212036108
#>  [6,]  0.811118929  0.2261341  0.1821841  0.96456971 -1.01640210 -0.533410636
#>  [7,] -0.735002273 -0.5610931 -1.2074401  0.75619792 -0.15845325  0.002205478
#>  [8,] -0.104626876 -0.5480882  1.5545471 -0.27871042 -2.25195394 -0.346823573
#>  [9,] -0.695506182  0.3304972 -1.7377733 -1.05027468 -0.22329538 -0.485861799
#> [10,]  0.003144412 -0.7534743 -0.3793331 -1.11388494 -0.76745061  0.216737308

X[] <- 1:100
X[, 1]

#>  [1]  1  2  3  4  5  6  7  8  9 10

X[1, ]  # not recommended for large matrices

#>  [1]  1 11 21 31 41 51 61 71 81 91

X[, -1]

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9]
#>  [1,]   11   21   31   41   51   61   71   81   91
#>  [2,]   12   22   32   42   52   62   72   82   92
#>  [3,]   13   23   33   43   53   63   73   83   93
#>  [4,]   14   24   34   44   54   64   74   84   94
#>  [5,]   15   25   35   45   55   65   75   85   95
#>  [6,]   16   26   36   46   56   66   76   86   96
#>  [7,]   17   27   37   47   57   67   77   87   97
#>  [8,]   18   28   38   48   58   68   78   88   98
#>  [9,]   19   29   39   49   59   69   79   89   99
#> [10,]   20   30   40   50   60   70   80   90  100

X[, c(TRUE, FALSE)]

#>       [,1] [,2] [,3] [,4] [,5]
#>  [1,]    1   21   41   61   81
#>  [2,]    2   22   42   62   82
#>  [3,]    3   23   43   63   83
#>  [4,]    4   24   44   64   84
#>  [5,]    5   25   45   65   85
#>  [6,]    6   26   46   66   86
#>  [7,]    7   27   47   67   87
#>  [8,]    8   28   48   68   88
#>  [9,]    9   29   49   69   89
#> [10,]   10   30   50   70   90

X[cbind(1:10, 1:10)] <- NA_real_
X[]

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   NA   11   21   31   41   51   61   71   81    91
#>  [2,]    2   NA   22   32   42   52   62   72   82    92
#>  [3,]    3   13   NA   33   43   53   63   73   83    93
#>  [4,]    4   14   24   NA   44   54   64   74   84    94
#>  [5,]    5   15   25   35   NA   55   65   75   85    95
#>  [6,]    6   16   26   36   46   NA   66   76   86    96
#>  [7,]    7   17   27   37   47   57   NA   77   87    97
#>  [8,]    8   18   28   38   48   58   68   NA   88    98
#>  [9,]    9   19   29   39   49   59   69   79   NA    99
#> [10,]   10   20   30   40   50   60   70   80   90    NA