Class FBM.code256

A reference class for storing and accessing up to 256 arbitrary different values using a Filebacked Big Matrix of type unsigned char. Compared to a Filebacked Big Matrix, it adds a slot code which is used as a lookup table of size 256.

FBM.code256(nrow, ncol, code = rep(NA_real_, 256), init = NULL,
  backingfile = tempfile(), create_bk = TRUE)

add_code256(x, code)

Arguments

nrow
Number of rows.

ncol
Number of columns.

code
A numeric vector (of length 256). You should construct it with rep(NA_real_, 256) and then replace the values which are of interest for you.

init
Either a single value (e.g. 0) or as many value as the number of elements of the FBM. Default doesn't initialize the matrix.

backingfile
Path to the file storing the Big Matrix on disk. An extension ".bk" will be automatically added. Default stores in the temporary directory.

create_bk
Create a backingfile (the default) or use an existing one (which should be named by the backingfile parameter and have an extension ".bk"). For example, this could be used to convert a filebacked big.matrix from package bigmemory to a FBM.

x
A FBM.

Examples

X <- FBM(10, 10, type = "raw")
X[] <- sample(as.raw(0:3), size = length(X), replace = TRUE)
X[]

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]   00   03   03   02   03   03   03   03   03    01
#>  [2,]   00   01   03   03   03   03   02   02   01    02
#>  [3,]   02   02   00   01   00   02   02   01   02    03
#>  [4,]   01   00   03   01   03   02   03   02   03    01
#>  [5,]   03   01   01   03   02   01   03   01   03    01
#>  [6,]   00   01   01   03   02   01   02   01   02    00
#>  [7,]   01   02   00   00   02   02   02   00   00    01
#>  [8,]   02   00   02   02   03   00   01   02   03    01
#>  [9,]   02   03   03   01   01   03   00   00   00    03
#> [10,]   00   01   03   01   01   03   01   00   02    01

code <- rep(NA_real_, 256)
code[1:3] <- c(1, 3, 5)

X.code <- add_code256(X, code)
X.code[]

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    1   NA   NA    5   NA   NA   NA   NA   NA     3
#>  [2,]    1    3   NA   NA   NA   NA    5    5    3     5
#>  [3,]    5    5    1    3    1    5    5    3    5    NA
#>  [4,]    3    1   NA    3   NA    5   NA    5   NA     3
#>  [5,]   NA    3    3   NA    5    3   NA    3   NA     3
#>  [6,]    1    3    3   NA    5    3    5    3    5     1
#>  [7,]    3    5    1    1    5    5    5    1    1     3
#>  [8,]    5    1    5    5   NA    1    3    5   NA     3
#>  [9,]    5   NA   NA    3    3   NA    1    1    1    NA
#> [10,]    1    3   NA    3    3   NA    3    1    5     3

# Or directly
X.code2 <- FBM.code256(10, 10, code, init = sample(as.raw(0:3), 100, TRUE))
X.code2[]

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    5   NA    5    3    5   NA    3    3    3    NA
#>  [2,]    5    3    1   NA    3    5    3    5    5     5
#>  [3,]    5    1    1    3   NA    1   NA    5    1    NA
#>  [4,]    3    5   NA   NA   NA    3    5    5   NA     5
#>  [5,]   NA    3    3    5   NA    1   NA    1    5     1
#>  [6,]    3   NA    3    5    5   NA    1    5    1     1
#>  [7,]    5    5    5    3   NA    5    3    1    1    NA
#>  [8,]    3    5    1    5    1    3    3    1    5     3
#>  [9,]    1   NA    5    1    1   NA    5   NA    5    NA
#> [10,]   NA    3    1    3   NA    3   NA    1    1     5

# Copy the FBM with another code
X.code3 <- X.code$copy(code = rnorm(256))
stopifnot(all.equal(X.code$code256, code))