Cross-Model Selection and Averaging.

This function separates the training set in K folds.

In turn,

each fold is considered as an inner validation set and the others (K - 1) folds form an inner training set,

the model is trained on the inner training set and the corresponding predictions (scores) for the inner validation set are computed,

the vector of coefficients corresponding to the vector of scores which maximizes feval is chosen.

The K resulting vectors of coefficients are then combined into one vector (see the method parameter).

big_CMSA(FUN, feval, X., y.train, ind.train = rows_along(X.),
  covar.train = NULL, K = 10, method = c("geometric-median", "mean-wise",
  "median-wise"), block.size = 1000, ncores = 1, ...)

Arguments

FUN

Function that computes a linear scores for different values along a regularization path. For now, this is relevant to big_spLinReg, big_spLogReg and big_spSVM. The corresponding <FUN>.predict function should exists.

feval

Custimized evaluation function. You should always aim at maximizing this function. If feval is like a loss function (which you want to minimize), you should use its opposite instead. Its only two arguments should be pred and target (e.g. AUC).

X.

Either a big.matrix or a big.matrix.descriptor.

y.train

Vector of responses, corresponding to ind.train.

ind.train

An optional vector of the row indices that are used, for the training part. If not specified, all rows are used. Don't use negative indices.

covar.train

Matrix of covariables to be added in each model to correct for confounders (e.g. the scores of PCA), corresponding to ind.train. Default is NULL and corresponds to only adding an intercept to each model.

K

Number of folds that are used.

method

Method for combining vectors of coefficients. The default uses the https://en.wikipedia.org/wiki/Geometric_median.

block.size

Maximum number of columns read at once. Default is 1000.

ncores

Number of cores used. Default doesn't use parallelism.

...

Extra parameters to be passed to FUN.

Value

A vector of resulting coefficients (intercept excluded), which corresponds to the training set. You might want to recompute an optimal intercept if needed.