Standard univariate statistics

Standard univariate statistics for columns of a big.matrix. For now, the sum and var are implemented (the mean and sd can easily be deduced, see examples).

big_colstats(X, ind.train = seq(nrow(X)), ind.col = seq(ncol(X)))

Arguments

X

A big.matrix. You shouldn't have missing values in your data.

ind.train

An optional vector of the row indices that are used, for the training part. If not specified, all rows are used.

ind.col

An optional vector of the column indices that are used. If not specified, all columns are used.

Value

Data.frame of two numeric vectors sum and var with the corresponding column statistics.

See also

colSums apply

Examples

# Simulating some data
X <- big.matrix(41, 7)
X[] <- rnorm(length(X))

# Check the results
print(test <- big_colstats(X))

#>         sum       var
#> 1 -2.077357 1.2049367
#> 2  1.261039 0.8672623
#> 3 -9.458770 0.8151757
#> 4 12.130566 1.0229330
#> 5  2.710006 1.2582299
#> 6 -8.844654 0.9613270
#> 7  2.526466 0.6622671

# Only with the first 10 rows
ind <- 1:10
print(test <- big_colstats(X, ind))

#>           sum       var
#> 1  2.69195033 1.0160267
#> 2  0.50652224 0.8966887
#> 3 -2.79466993 0.2228012
#> 4  0.02753026 0.5934445
#> 5  2.32821278 0.8872320
#> 6  0.79318123 1.1819570
#> 7  0.02530883 0.7607667

print(all.equal(test$sum, colSums(X[ind, ])))

#> [1] TRUE

print(all.equal(test$var, apply(X[ind, ], 2, var)))

#> [1] TRUE

# deduce mean and sd
# note that the are also implemented in big_scale()
means <- test$sum / length(ind) # if using all rows,
                                # divide by nrow(X) instead
print(all.equal(means, colMeans(X[ind, ])))

#> [1] TRUE

sds <- sqrt(test$var)
print(all.equal(sds, apply(X[ind, ], 2, sd)))

#> [1] TRUE

# NA in, NA out
X[2, 5] <- NA
print(big_colstats(X))

#>         sum       var
#> 1 -2.077357 1.2049367
#> 2  1.261039 0.8672623
#> 3 -9.458770 0.8151757
#> 4 12.130566 1.0229330
#> 5        NA        NA
#> 6 -8.844654 0.9613270
#> 7  2.526466 0.6622671