Counts

Counts by columns (or rows) the number of each unique element of a BM.code.

big_counts(X.code, ind.row = rows_along(X.code),
  ind.col = cols_along(X.code), byrow = FALSE)

Arguments

X.code
Either a BM.code or a BM.code.descriptor.

ind.row
An optional vector of the row indices that are used. If not specified, all rows are used. Don't use negative indices.

ind.col
An optional vector of the column indices that are used. If not specified, all columns are used. Don't use negative indices.

byrow
Count by rows rather than columns? Default is FALSE.

Value

A matrix of counts of K x m (or n) elements, where

K is the number of unique elements of the BM.code,

n is its number of rows,

m is its number of columns.

Beware that K is up to 256. So, if you apply this on a big.matrix of one million columns, you will create a matrix of nearly 1Gb!.

Examples

set.seed(11)

a <- matrix(as.raw(0), 14, 11)
a[] <- sample(as.raw(0:3), size = length(a), replace = TRUE)
X <- as.big.matrix(a)
X[,]

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11]
#>  [1,]   01   02   00   01   01   02   03   02   01    00    02
#>  [2,]   00   02   01   02   03   01   01   00   03    03    02
#>  [3,]   02   01   02   01   00   01   00   03   02    02    01
#>  [4,]   00   01   01   03   02   00   02   02   01    02    02
#>  [5,]   00   00   01   02   00   03   03   01   03    03    00
#>  [6,]   03   01   00   00   03   02   00   01   01    02    02
#>  [7,]   00   00   03   02   00   01   01   00   02    00    02
#>  [8,]   01   02   02   03   02   01   03   01   00    02    01
#>  [9,]   03   01   01   02   03   01   00   02   00    00    03
#> [10,]   00   01   00   02   00   00   01   02   01    01    01
#> [11,]   00   00   01   01   00   00   03   02   03    01    02
#> [12,]   01   01   00   01   01   01   00   00   00    02    03
#> [13,]   03   01   00   00   00   02   03   02   01    03    00
#> [14,]   03   00   00   01   02   00   00   00   00    03    03

code <- rep(NA_real_, 256)
code[1:4] <- c(2, 5, 9, NA)
X2 <- as.BM.code(X, code)
X2[,]

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11]
#>  [1,]    5    9    2    5    5    9   NA    9    5     2     9
#>  [2,]    2    9    5    9   NA    5    5    2   NA    NA     9
#>  [3,]    9    5    9    5    2    5    2   NA    9     9     5
#>  [4,]    2    5    5   NA    9    2    9    9    5     9     9
#>  [5,]    2    2    5    9    2   NA   NA    5   NA    NA     2
#>  [6,]   NA    5    2    2   NA    9    2    5    5     9     9
#>  [7,]    2    2   NA    9    2    5    5    2    9     2     9
#>  [8,]    5    9    9   NA    9    5   NA    5    2     9     5
#>  [9,]   NA    5    5    9   NA    5    2    9    2     2    NA
#> [10,]    2    5    2    9    2    2    5    9    5     5     5
#> [11,]    2    2    5    5    2    2   NA    9   NA     5     9
#> [12,]    5    5    2    5    5    5    2    2    2     9    NA
#> [13,]   NA    5    2    2    2    9   NA    9    5    NA     2
#> [14,]   NA    2    2    5    9    2    2    2    2    NA    NA

# by columns
big_counts(X2)

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11]
#> 2       6    4    6    2    6    4    5    4    4     3     2
#> 5       3    7    5    5    2    6    3    3    5     2     3
#> 9       1    3    2    5    3    3    1    6    2     5     6
#> <NA>    4    0    1    2    3    1    5    1    3     4     3

apply(X2[,], 2, table, exclude = NULL)

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11]
#> 2       6    4    6    2    6    4    5    4    4     3     2
#> 5       3    7    5    5    2    6    3    3    5     2     3
#> 9       1    3    2    5    3    3    1    6    2     5     6
#> <NA>    4    0    1    2    3    1    5    1    3     4     3

# by rows
big_counts(X2, byrow = TRUE)

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> 2       2    2    2    2    4    3    5    1    3     4     4     4     4     6
#> 5       4    3    4    3    2    3    2    4    3     5     3     5     2     1
#> 9       4    3    4    5    1    3    3    4    2     2     2     1     2     1
#> <NA>    1    3    1    1    4    2    1    2    3     0     2     1     3     3

apply(X2[,], 1, table, exclude = NULL)

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> 2       2    2    2    2    4    3    5    1    3     4     4     4     4     6
#> 5       4    3    4    3    2    3    2    4    3     5     3     5     2     1
#> 9       4    3    4    5    1    3    3    4    2     2     2     1     2     1
#> <NA>    1    3    1    1    4    2    1    2    3     0     2     1     3     3