Counts

Counts by columns (or rows) the number of each unique element of a FBM.code256.

big_counts(X.code, ind.row = rows_along(X.code),
  ind.col = cols_along(X.code), byrow = FALSE)

Arguments

X.code
A FBM.code256.

ind.row
An optional vector of the row indices that are used. If not specified, all rows are used. Don't use negative indices.

ind.col
An optional vector of the column indices that are used. If not specified, all columns are used. Don't use negative indices.

byrow
Count by rows rather than columns? Default is FALSE (columns).

Value

A matrix of counts of K x m (or n) elements, where

K is the number of unique elements of the BM.code,

n is its number of rows,

m is its number of columns.

Beware that K is up to 256. So, if you apply this on a Filebacked Big Matrix of one million columns, you will create a matrix of nearly 1GB!.

Examples

X <- big_attachExtdata()
class(X)

#> [1] "FBM.code256"
#> attr(,"package")
#> [1] "bigstatsr"

X[1:5, 1:10]

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,]    2    2    0    2    1    1    2    0    1     0
#> [2,]    1    2    1    2    2    1    2    0    0     1
#> [3,]    2    1    1    2    0    1    2    0    0     2
#> [4,]    2    2    0    2    0    1    2    0    2     1
#> [5,]    1    2    2    2    0    0    1    2    1     1

# Without the "decoding"
X2 <- big_copy(X)
class(X2)

#> [1] "FBM"
#> attr(,"package")
#> [1] "bigstatsr"

X2[1:5, 1:10]

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,]   02   02   00   02   01   01   02   00   01    00
#> [2,]   01   02   01   02   02   01   02   00   00    01
#> [3,]   02   01   01   02   00   01   02   00   00    02
#> [4,]   02   02   00   02   00   01   02   00   02    01
#> [5,]   01   02   02   02   00   00   01   02   01    01

# Change the code
code <- rep(NA_real_, 256)
code[1:3] <- c(2, 5, 9)
X$code256 <- code
X[1:5, 1:10]

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,]    9    9    2    9    5    5    9    2    5     2
#> [2,]    5    9    5    9    9    5    9    2    2     5
#> [3,]    9    5    5    9    2    5    9    2    2     9
#> [4,]    9    9    2    9    2    5    9    2    9     5
#> [5,]    5    9    9    9    2    2    5    9    5     5

# by columns
big_counts(X, ind.col = 1:10)

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> 2      63   21   37   20  113   66   10  114  124   123
#> 5     228  171  171  151  246  236  112  269  250   235
#> 9     226  325  309  346  158  215  395  134  143   159
#> <NA>    0    0    0    0    0    0    0    0    0     0

apply(X[, 1:10], 2, table, exclude = NULL)

#>   [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> 2   63   21   37   20  113   66   10  114  124   123
#> 5  228  171  171  151  246  236  112  269  250   235
#> 9  226  325  309  346  158  215  395  134  143   159

# by rows
big_counts(X, ind.row = 1:10, byrow = TRUE)

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> 2     532  568  565  540  575  564  541  506  525   510
#> 5    1777 1780 1712 1755 1722 1752 1765 1805 1758  1831
#> 9    2233 2194 2265 2247 2245 2226 2236 2231 2259  2201
#> <NA>    0    0    0    0    0    0    0    0    0     0

apply(X[1:10, ], 1, table, exclude = NULL)

#>   [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> 2  532  568  565  540  575  564  541  506  525   510
#> 5 1777 1780 1712 1755 1722 1752 1765 1805 1758  1831
#> 9 2233 2194 2265 2247 2245 2226 2236 2231 2259  2201