Cross-product with a vector

Cross-product between a "big.matrix" and a vector.

big_cprodVec(X., y, ind.row = rows_along(X.), ind.col = cols_along(X.))

Arguments

X.

Either a big.matrix or a big.matrix.descriptor.

y

A vector of same size as ind.row.

ind.row

An optional vector of the row indices that are used. If not specified, all rows are used. Don't use negative indices.

ind.col

An optional vector of the column indices that are used. If not specified, all columns are used. Don't use negative indices.

Value

\(X^T \cdot y\).

Examples

X.desc <- big_attachExtdata()
n <- nrow(X.desc)
m <- ncol(X.desc)
y <- rnorm(n)

test <- big_cprodVec(X.desc, y)             # vector
true <- crossprod(attach.BM(X.desc)[,], y)  # one-column matrix
all.equal(test, as.numeric(true))

#> [1] TRUE

# subsetting
ind.row <- sample(n, n/2)
ind.col <- sample(m, m/2)

## Not run: test2 &lt;- big_cprodVec(X.desc, y, ind.row, ind.col)
# returns an error. You need to use the subset of y:
test2 <- big_cprodVec(X.desc, y[ind.row], ind.row, ind.col)
true2 <- crossprod(attach.BM(X.desc)[ind.row, ind.col], y[ind.row])
all.equal(test2, as.numeric(true2))

#> [1] TRUE