Crossprod

Compute \(X.row^T X.row\) for a Filebacked Big Matrix X after applying a particular scaling to it.

big_crossprodSelf(
  X,
  fun.scaling = big_scale(center = FALSE, scale = FALSE),
  ind.row = rows_along(X),
  ind.col = cols_along(X),
  block.size = block_size(nrow(X))
)

# S4 method for FBM,missing
crossprod(x, y)

Arguments

X
A FBM.

fun.scaling
A function that returns a named list of mean and sd for every column, to scale each of their elements such as followed: $$\frac{X_{i,j} - mean_j}{sd_j}.$$ Default doesn't use any scaling.

ind.row
An optional vector of the row indices that are used. If not specified, all rows are used. Don't use negative indices.

ind.col
An optional vector of the column indices that are used. If not specified, all columns are used. Don't use negative indices.

block.size
Maximum number of columns read at once. Default uses block_size.

x
A 'double' FBM.

y
Missing.

Value

A temporary FBM, with the following two attributes:

a numeric vector center of column scaling,

a numeric vector scale of column scaling.

Matrix parallelization

Large matrix computations (crossprods) are made block-wise and won't be parallelized in order to not have to reduce the size of these blocks. Instead, you may use Microsoft R Open in order to accelerate these block matrix computations.

See also

Examples

X <- FBM(13, 17, init = rnorm(221))
true <- crossprod(X[])

# No scaling
K1 <- crossprod(X)
class(K1)

#> [1] "matrix"

stopifnot(all.equal(K1, true))

K2 <- big_crossprodSelf(X)
class(K2)

#> [1] "FBM"
#> attr(,"package")
#> [1] "bigstatsr"

K2$backingfile

#> [1] "/tmp/RtmpUEgzDk/file506053fa6749.bk"

stopifnot(all.equal(K2[], true))

microbenchmark::microbenchmark(
  crossprod(X[]), crossprod(X), big_crossprodSelf(X)
)

#> Unit: microseconds
#>                  expr      min        lq      mean   median        uq
#>        crossprod(X[])  149.047  171.3390  197.9047  184.394  194.6110
#>          crossprod(X)   88.187  104.7775  128.0225  116.301  130.1495
#>  big_crossprodSelf(X) 3465.919 3903.4415 5591.8433 4118.097 4430.2530
#>         max neval
#>     485.534   100
#>     288.712   100
#>  135426.034   100

# big_crossprodSelf() provides some scaling and subsetting
# Example using only half of the data:
n <- nrow(X)
ind <- sort(sample(n, n/2))
K3 <- big_crossprodSelf(X, fun.scaling = big_scale(), ind.row = ind)
true2 <- crossprod(scale(X[ind, ]))
stopifnot(all.equal(K3[], true2))