Crossprod

Compute \(X.row^T X.row\) for a big.matrix X after applying a particular scaling to it.

big_crossprodSelf(X., fun.scaling, ind.row = rows_along(X.),
  block.size = 1000)

Arguments

X.

Either a big.matrix or a big.matrix.descriptor.

fun.scaling

A function that returns a named list of mean and sd for every column, to scale each of their elements such as followed: $$\frac{X_{i,j} - mean_j}{sd_j}$$.

ind.row

An optional vector of the row indices that are used. If not specified, all rows are used. Don't use negative indices.

block.size

Maximum number of columns read at once. Default is 1000.

Value

A list of

\(K = X.row^T X.row\),

a numeric vector mean of column scaling,

a numeric vector sd of column scaling.

See also

crossprod

Examples

X.desc <- big_attachExtdata()

# Comparing with tcrossprod
big_noscale <- big_scale(center = FALSE)
test <- big_crossprodSelf(X.desc, fun.scaling = big_noscale)
dim(test$K)

#> [1] 4542 4542

true <- crossprod(attach.BM(X.desc)[,])
all.equal(test$K, true)

#> [1] TRUE

# Using only half of the data
n <- nrow(X.desc)
ind <- sort(sample(n, n/2))
test2 <- big_crossprodSelf(X.desc, fun.scaling = big_noscale,
                           ind.row = ind)
dim(test2$K)

#> [1] 4542 4542

true2 <- crossprod(attach.BM(X.desc)[ind, ])
all.equal(test2$K, true2)

#> [1] TRUE