Split-parApply-Combine

A Split-Apply-Combine strategy to parallelize the evaluation of a function.

big_parallelize(X., p.FUN, p.combine, ncores, ind = cols_along(X.), ...)

Arguments

X.
Either a big.matrix or a big.matrix.descriptor.

p.FUN
The function to be applied. It must take a big.matrix.descriptor as first argument and provide some arguments for subsetting.

p.combine
function that is used by foreach to process the tasks results as they generated. This can be specified as either a function or a non-empty character string naming the function. Specifying 'c' is useful for concatenating the results into a vector, for example. The values 'cbind' and 'rbind' can combine vectors into a matrix. The values '+' and '*' can be used to process numeric data. By default, the results are returned in a list.

ncores
Number of cores used. Default doesn't use parallelism. We advise you to use only half of the cores you have.

ind
Initial vector of subsetting indices. Default is the vector of all column indices.

...
Extra arguments to be passed to p.FUN.

Value

The result of foreach.

Details

This function splits indices in parts, then apply a given function to each part and finally combine the results.

See also

big_apply

Examples

X.desc <- big_attachExtdata()

### Computation on all the matrix
true <- big_colstats(X.desc)

big_colstats_sub <- function(X.desc, ind, cols) {
  big_colstats(X.desc, ind.col = cols[ind])
}
# 1. the computation is split along all the columns
# 2. for each part the computation is done, using `big_colstats`
# 3. the results (data.frames) are combined via `rbind`.
test <- big_parallelize(X.desc, p.FUN = big_colstats_sub,
                        p.combine = 'rbind', ncores = 2,
                        cols = cols_along(X.desc))
all.equal(test, true)

#> [1] TRUE

### Computation on a part of the matrix
n <- nrow(X.desc)
m <- ncol(X.desc)
rows <- sort(sample(n, n/2)) # sort to provide some locality in accesses
cols <- sort(sample(m, m/2)) # idem

true2 <- big_colstats(X.desc, ind.row = rows, ind.col = cols)

big_colstats_sub2 <- function(X.desc, ind, rows, cols) {
  big_colstats(X.desc, ind.row = rows, ind.col = cols[ind])
}
# This doesn't work because, by default, the computation is spread
# along all columns. We must explictly specify the `ind` parameter.
not_run({
  test2 <- big_parallelize(X.desc, p.FUN = big_colstats_sub2,                                    p.combine = 'rbind', ncores = 2,                                    rows = rows, cols = cols)
})

# This now works, using `ind = seq_along(cols)`.
test2 <- big_parallelize(X.desc, p.FUN = big_colstats_sub2,
                         p.combine = 'rbind', ncores = 2,
                         ind = seq_along(cols),
                         rows = rows, cols = cols)
all.equal(test2, true2)

#> [1] TRUE