Read a file

Read a file as a Filebacked Big Matrix by using package bigreadr. For a mini-tutorial, please see this vignette.

big_read(file, select, nrow, type = c("double", "integer",
  "unsigned short", "unsigned char", "raw"),
  backingfile = drop_ext(file), nb_parts = NULL, skip = 0,
  progress = TRUE, ...)

Arguments

file
Path to file that you want to read.

select
Indices of columns to read (sorted). The length of select will be the number of columns of the resulting FBM.

nrow
Number of rows of the resulting FBM. This is basically (nlines - skip - header).

type
Type of the Filebacked Big Matrix (default is double). Either

"double"

"integer"

"unsigned short": can store integer values from 0 to 65535. It has vocation to become the basis for a FBM.code65536 class for accessing strings.

"raw" or "unsigned char": can store integer values from 0 to 255. It is the basis for the FBM.code256 class for accessing 256 arbitrary different numeric values. It is used in package bigsnpr.

backingfile
Path to the file storing the Big Matrix on disk. An extension ".bk" will be automatically added. Default uses file without its extension.

nb_parts
Number of parts in which to split reading. Parts are referring to blocks of selected columns. Default use global option bigstatsr.block.sizeGB to set a good value.

skip
Number of lines to skip at the beginning of file.

progress
Show progress? Default is TRUE.

...
More arguments to be passed to data.table::fread.

Value

A Filebacked Big Matrix of type '<type>' with <nrow> rows and <length(select)> columns.