Some scaling functions

Some scaling functions for a big.matrix to be used as the fun.scaling parameter of some functions of this package.

big_scale(center = TRUE, scale = TRUE)

Arguments

center

A logical value: whether to return means or 0s.

scale

A logical value: whether to return sds or 1s. You can't use scale without using center.

Value

A new function that returns a named list of two vectors "mean" and "sd" which are as long as the number of columns of X.

See also

scale

Examples

# Simulating some data
X <- big.matrix(41, 17)
X[] <- rnorm(length(X))

# No scaling
big_noscale <- big_scale(center = FALSE, scale = FALSE)
print(class(big_noscale)) # big_scale returns a new function

#> [1] "function"

print(big_noscale(X))

#> $mean
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $sd
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 

big_noscale2 <- big_scale(center = FALSE)
print(big_noscale2(X)) # you can't scale without centering

#> $mean
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $sd
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 

# Centering
big_center <- big_scale(scale = FALSE)
print(big_center(X))

#> $mean
#>  [1] -0.07376648 -0.25932683 -0.19646230 -0.16035114  0.21224444 -0.48983288
#>  [7] -0.08434935  0.04082948  0.22828540  0.14248646  0.18224363 -0.22549647
#> [13] -0.04449430  0.25031149  0.01612221  0.18043020  0.18428189
#> 
#> $sd
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 

# + scaling
print(test <- big_scale()(X, 1:5)) # note that big_scale() is a function

#> $mean
#>  [1] -0.008564808  0.290281873 -0.364288843 -0.137087174  0.002599303
#>  [6] -0.516994734 -0.241953769  0.753523640  0.225174394 -0.054998260
#> [11]  0.277181288 -0.108770338 -0.249249598  0.147172818  0.255645720
#> [16] -0.155503995 -0.162564338
#> 
#> $sd
#>  [1] 0.6596598 1.1225398 0.6108999 1.1067161 1.2890436 1.1004375 0.9998248
#>  [8] 0.4695924 0.6416575 1.2021877 0.8500464 0.5461798 0.8646184 1.2226307
#> [15] 0.8729338 1.0623704 1.4852248
#> 

print(all.equal(test$mean, colMeans(X[1:5, ])))

#> [1] TRUE

print(all.equal(test$sd, apply(X[1:5, ], 2, sd)))

#> [1] TRUE