Some scaling functions

Some scaling functions for a big.matrix to be used as the fun.scaling parameter of some functions of this package.
big_scale(center = TRUE, scale = TRUE)

Arguments

center 	A logical value: whether to return means or 0s.

scale 	A logical value: whether to return sds or 1s. You can't use scale without using center.

Value
A new function that returns a data.frame of two vectors "mean" and "sd" which are of the length of ind.col.
Details
One could think about less common scalings, such as for example the "y-aware" scaling which uses the inverse of betas of column-wise linear regression as scaling. See https://goo.gl/8G8WMa for details. It would be easy to implement it using big_colstats to get column means and big_univLinReg to get betas (and then inverse them).
See also
scale
Examples

X.desc <- big_attachExtdata()

# No scaling
big_noscale <- big_scale(center = FALSE, scale = FALSE)
class(big_noscale) # big_scale returns a new function

#> [1] "function"

str(big_noscale(X.desc))

#> 'data.frame':	4542 obs. of  2 variables:
#>  $ mean: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ sd  : num  1 1 1 1 1 1 1 1 1 1 ...

big_noscale2 <- big_scale(center = FALSE)
str(big_noscale2(X.desc)) # you can't scale without centering

#> 'data.frame':	4542 obs. of  2 variables:
#>  $ mean: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ sd  : num  1 1 1 1 1 1 1 1 1 1 ...

# Centering
big_center <- big_scale(scale = FALSE)
str(big_center(X.desc))

#> 'data.frame':	4542 obs. of  2 variables:
#>  $ mean: num  1.32 1.59 1.53 1.63 1.09 ...
#>  $ sd  : num  1 1 1 1 1 1 1 1 1 1 ...

# + scaling
str(big_scale()(X.desc))

#> 'data.frame':	4542 obs. of  2 variables:
#>  $ mean: num  1.32 1.59 1.53 1.63 1.09 ...
#>  $ sd  : num  0.679 0.569 0.627 0.558 0.719 ...