Some scaling functions

Some scaling functions for a big.matrix to be used as the fun.scaling parameter of some functions of this package.

big_scale(center = TRUE, scale = TRUE)

Arguments

center

A logical value: whether to return means or 0s.

scale

A logical value: whether to return sds or 1s. You can't use scale without using center.

Value

A new function that returns a named list of two vectors "mean" and "sd" which are as long as the number of columns of X.

See also

scale

Examples

# Simulating some data
X <- big.matrix(41, 17)
X[] <- rnorm(length(X))

# No scaling
big_noscale <- big_scale(center = FALSE, scale = FALSE)
print(class(big_noscale)) # big_scale returns a new function

#> [1] "function"

print(big_noscale(X))

#> $mean
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $sd
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 

big_noscale2 <- big_scale(center = FALSE)
print(big_noscale2(X)) # you can't scale without centering

#> $mean
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $sd
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 

# Centering
big_center <- big_scale(scale = FALSE)
print(big_center(X))

#> $mean
#>  [1]  0.19728549  0.10618530 -0.27338724 -0.09691088 -0.13808392  0.04508504
#>  [7]  0.06391175 -0.11578813  0.19390636  0.24291342  0.08676105  0.28331784
#> [13] -0.07894416  0.05404656 -0.06758716  0.12365584  0.16139489
#> 
#> $sd
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 

# + scaling
print(test <- big_scale()(X, 1:5)) # note that big_scale() is a function

#> $mean
#>  [1] -0.5174596674 -0.1338099291 -0.5232721700  0.2962544024 -0.3672392842
#>  [6]  0.0675924298  0.8255243377  0.2882485619  0.3470733265 -0.0473455983
#> [11] -0.7222461341 -0.0009355725  0.5999761984  0.0575834871 -0.0002809985
#> [16]  0.0168633960  0.1431180810
#> 
#> $sd
#>  [1] 1.2237682 1.4010876 1.3655102 0.5688388 1.2492848 1.4240795 1.2072633
#>  [8] 1.1448961 0.4985260 1.2459639 1.1500712 0.4539137 0.6649315 0.6409716
#> [15] 0.9785205 1.1079142 0.6501422
#> 

print(all.equal(test$mean, colMeans(X[1:5, ])))

#> [1] TRUE

print(all.equal(test$sd, apply(X[1:5, ], 2, sd)))

#> [1] TRUE