Sparse SVM

Fit solution paths for sparse linear SVM regularized by lasso or elastic-net over a grid of values for the regularization parameter lambda.

big_spSVM(X, y01.train, ind.train = rows_along(X), covar.train = NULL, ...)

Arguments

X
A FBM.

y01.train
Vector of responses, corresponding to ind.train. Must be only 0s and 1s.

ind.train
An optional vector of the row indices that are used, for the training part. If not specified, all rows are used. Don't use negative indices.

covar.train
Matrix of covariables to be added in each model to correct for confounders (e.g. the scores of PCA), corresponding to ind.train. Default is NULL and corresponds to only adding an intercept to each model.

...
Arguments passed on to COPY_sparseSVM

alpha

The elastic-net mixing parameter that controls the relative contribution from the lasso and the ridge penalty. It must be a number between 0 and 1. alpha=1 is the lasso penalty and alpha=0 the ridge penalty.

gamma

The tuning parameter for huberization smoothing of hinge loss. Default is 0.1.

nlambda

The number of lambda values. Default is 100.

lambda.min

The smallest value for lambda, as a fraction of lambda.max, the data derived entry value. Default is 0.01 if the number of observations is larger than the number of variables and 0.05 otherwise.

lambda

A user-specified sequence of lambda values. Typical usage is to leave blank and have the program automatically compute a lambda sequence based on nlambda and lambda.min. Specifying lambda overrides this. This argument should be used with care and supplied with a decreasing sequence instead of a single value.

screen

Screening rule to be applied at each lambda that discards variables for speed. Either "ASR" (default), "SR" or "none". "SR" stands for the strong rule, and "ASR" for the adaptive strong rule. Using "ASR" typically requires fewer iterations to converge than "SR", but the computing time are generally close. Note that the option "none" is used mainly for debugging, which may lead to much longer computing time.

max.iter

Maximum number of iterations. Default is 1000.

eps

Convergence threshold. The algorithms continue until the maximum change in the objective after any coefficient update is less than eps times the null deviance. Default is 1E-7.

dfmax

Upper bound for the number of nonzero coefficients. The algorithm exits and returns a partial path if dfmax is reached. Useful for very large dimensions.

penalty.factor

A numeric vector of length equal to the number of variables. Each component multiplies lambda to allow differential penalization. Can be 0 for some variables, in which case the variable is always in the model without penalization. Default is 1 for all variables.

message

If set to TRUE, sparseSVM will inform the user of its progress. Default is FALSE.

Value

A named list containing:

call

The call that produced this object.

intercept

A vector of intercepts, corresponding to each lambda.

beta

The fitted matrix of coefficients. The number of rows is equal to the number of coefficients, and the number of columns is equal to nlambda.

iter

A vector of length nlambda containing the number of iterations until convergence at each value of lambda.

saturated

A logical flag for whether the number of nonzero coefficients has reached dfmax.

lambda

The sequence of regularization parameter values in the path.

alpha

Input parameter.

gamma

Input parameter.

penalty.factor

Input parameter.

levels

Levels of the output class labels.

Details

This is a modified version of one function of package sparseSVM. This algorithm uses semi-newton coordinate descent. There may exist some faster algorithm. For now, this function is not exported anymore. This is because it doesn't match the results from package sparseSVM anymore, due to a change the author made. I've contacted him multiple times, without success. If you really want to use this function, use bigstatsr:::big_spSVM.

See also

LiblineaR sparseSVM

Examples

set.seed(1)

big_spSVM <- bigstatsr:::big_spSVM

# simulating some data
N <- 73
M <- 430
X <- FBM(N, M, init = rnorm(N * M, sd = 5))
y <- sample(0:1, size = N, replace = TRUE)
y.factor <- factor(y, levels = c(1, 0))
covar <- matrix(rnorm(N * 3), N)

# error, only handle standard R matrices
# NOT RUN {
test <- sparseSVM::sparseSVM(X, y)
# }

# OK here
test2 <- big_spSVM(X, y)
str(test2)

#> List of 10
#>  $ call          : language COPY_sparseSVM(X = X, y.train = y01.train, ind.train = ind.train, covar.train = covar.train)
#>  $ intercept     : Named num [1:100] -1 -0.886 -0.867 -0.275 -0.26 ...
#>   ..- attr(*, "names")= chr [1:100] "0.2624" "0.2545" "0.2469" "0.2396" ...
#>  $ beta          :Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   .. ..@ i       : int [1:10849] 183 201 304 341 9 183 193 201 304 341 ...
#>   .. ..@ p       : int [1:101] 0 0 4 10 19 29 41 55 77 100 ...
#>   .. ..@ Dim     : int [1:2] 430 100
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:100] "0.2624" "0.2545" "0.2469" "0.2396" ...
#>   .. ..@ x       : num [1:10849] 0.011 -0.00826 0.00404 0.00533 -0.00063 ...
#>   .. ..@ factors : list()
#>  $ iter          : int [1:100] 0 9 6 84 3 2 5 6 5 4 ...
#>  $ saturated     : logi FALSE
#>  $ lambda        : num [1:100] 0.262 0.255 0.247 0.24 0.232 ...
#>  $ alpha         : num 1
#>  $ gamma         : num 0.1
#>  $ penalty.factor: num [1:430] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ levels        : int [1:2] 0 1
#>  - attr(*, "class")= chr "big_sp"

# how to use covariables?
X2 <- cbind(X[], covar)
test <- sparseSVM::sparseSVM(X2, y.factor, alpha = 0.5)
test2 <- big_spSVM(X, y, covar.train = covar, alpha = 0.5)
# verification
all.equal(test2$lambda, test$lambda)

#> [1] TRUE

all.equal(as.matrix(test2$beta), test$weights[-1, ], check.attributes = FALSE)

#> [1] TRUE

all.equal(test2$intercept, test$weights[1, ])

#> [1] TRUE