Transposition

This function implements a simple cache-oblivious algorithm for the transposition of a Filebacked Big Matrix.

big_transpose(X, ...)

Arguments

X
A FBM.

...
Arguments passed on to FBM

backingfile

Path to the file storing the Big Matrix on disk. An extension ".bk" will be automatically added. Default stores in the temporary directory.

create_bk

Create a backingfile (the default) or use an existing one (which should be named by the backingfile parameter and have an extension ".bk"). For example, this could be used to convert a filebacked big.matrix from package bigmemory to a FBM.

save

Whether to save the result object in an ".rds" file alongside the backingfile. Default is FALSE.

Value

The new transposed Filebacked Big Matrix (or its descriptor). Its dimensions and type are automatically determined from the input Filebacked Big Matrix.

Examples

X <- FBM(10, 5, init = rnorm(50))
X[]

#>             [,1]       [,2]        [,3]       [,4]       [,5]
#>  [1,] -2.4298220  0.8026269  0.03442445  0.9604049 -1.1837680
#>  [2,]  0.6840947 -1.5066966  0.31560441 -1.1723163 -0.3105925
#>  [3,] -0.4401606  0.2411087  1.04898341 -0.2596051 -0.7069316
#>  [4,] -0.7617191  0.2417620  0.24066041  0.6155269 -0.9518207
#>  [5,]  0.1342995  1.6173050 -0.66118867 -1.9269711 -0.9555700
#>  [6,] -0.2323571  0.3694784 -0.94931860  0.7119835 -0.1826885
#>  [7,] -0.5040429 -0.5738322 -1.29325532 -0.5350299  0.5131919
#>  [8,]  1.2085269 -0.9312511  0.77767228  0.4241146 -0.3920569
#>  [9,]  1.3328910 -0.5952995 -0.93097554  1.6491252  2.0074487
#> [10,]  0.3425188 -0.3418070  0.94231203 -1.4270323  0.7090764

Xt <- big_transpose(X)
identical(t(X[]), Xt[])

#> [1] TRUE

X <- big_attachExtdata()
Xt <- big_transpose(X)
identical(t(X[]), Xt[])

#> [1] TRUE