Transposition

This function implements a simple cache-oblivious algorithm for the transposition of a "big.matrix".

big_transpose(X., descriptor = TRUE, ...)

Arguments

X.

Either a big.matrix or a big.matrix.descriptor.

descriptor

Return the descriptor of the transposed big.matrix or directly the big.matrix object? Default is TRUE (the descriptor).

...

Arguments passed on to bigmemory::big.matrix

separated

use separated column organization of the data; see details.

backingfile

the root name for the file(s) for the cache of x.

backingpath

the path to the directory containing the file backing cache.

descriptorfile

the name of the file to hold the backingfile description, for subsequent use with attach.big.matrix; if NULL, the backingfile is used as the root part of the descriptor file name. The descriptor file is placed in the same directory as the backing files.

binarydescriptor

the flag to specify if the binary RDS format should be used for the backingfile description, for subsequent use with attach.big.matrix; if NULL of FALSE, the dput() file format is used.

shared

TRUE by default, and always TRUE if the big.matrix is file-backed. For a non-filebacked big.matrix, shared=FALSE uses non-shared memory, which can be more stable for large (say, >50% of RAM) objects. Shared memory allocation can sometimes fail in such cases due to exhausted shared-memory resources in the system.

Value

The new transposed big.matrix (or its descriptor). Its dimension, type and dimnames are automatically determined from the input big.matrix.

Examples

X.desc <- big_attachExtdata()

tmpfile <- tempfile()
Xt.desc <- big_transpose(X.desc, descriptor = TRUE,
                         backingfile = basename(tmpfile),
                         backingpath = dirname(tmpfile),
                         descriptorfile = paste0(basename(tmpfile), ".desc"))

identical(t(attach.BM(X.desc)[,]), attach.BM(Xt.desc)[,])

#> [1] TRUE