Column-wise logistic regression

Slopes of column-wise logistic regressions of each column of a big.matrix, with some other associated statistics. Covariates can be added to correct for confounders.

big_univLogReg(X., y01.train, ind.train = rows_along(X.),
  ind.col = cols_along(X.), covar.train = NULL, ncores2 = 1,
  tol = 1e-08, maxiter = 20)

Arguments

X.

Either a big.matrix or a big.matrix.descriptor.

y01.train

Vector of responses, corresponding to ind.train. Must be 0s and 1s.

ind.train

An optional vector of the row indices that are used, for the training part. If not specified, all rows are used. Don't use negative indices.

ind.col

An optional vector of the column indices that are used. If not specified, all columns are used. Don't use negative indices.

covar.train

Matrix of covariables to be added in each model to correct for confounders (e.g. the scores of PCA), corresponding to ind.train. Default is NULL and corresponds to only adding an intercept to each model.

ncores2

Number of cores used. Default doesn't use parallelism. For this function, use only half of the cores you have.

tol

Relative tolerance to assess convergence of the coefficient. Default is 1e-8.

maxiter

Maximum number of iterations before giving up. Default is 20. Usually, convergence is reached within 3 or 4 iterations. If there is not convergence, glm is used instead for the corresponding column.

Value

A data.frame with as elements:

the slopes of each regression,

the standard errors of each slope,

the number of iteration for each slope. If is NA, this means that the algorithm didn't converge, and glm was used instead.

the z-scores associated with each slope,

the p-values associated with each z-score.

See also

glm

Examples

X.desc <- big_attachExtdata()
n <- nrow(X.desc)
y01 <- sample(0:1, size = n, replace = TRUE)
covar <- matrix(rnorm(n * 3), n)

X1 <- attach.BM(X.desc)[, 1] # only first column of the `big.matrix`

# Without covar
str(big_univLogReg(X.desc, y01))

#> 'data.frame':	4542 obs. of  5 variables:
#>  $ estim  : num  -0.0208 -0.1801 0.0419 -0.1379 0.1569 ...
#>  $ std.err: num  0.13 0.157 0.141 0.16 0.123 ...
#>  $ niter  : int  3 4 3 4 4 3 4 3 4 4 ...
#>  $ z.score: num  -0.16 -1.149 0.298 -0.864 1.273 ...
#>  $ p.value: num  0.873 0.251 0.766 0.388 0.203 ...

summary(glm(y01 ~ X1, family = "binomial"))$coefficients[2, ]

#>    Estimate  Std. Error     z value    Pr(>|z|) 
#> -0.02080582  0.13038263 -0.15957509  0.87321580 

# With all data
str(big_univLogReg(X.desc, y01, covar.train = covar))

#> 'data.frame':	4542 obs. of  5 variables:
#>  $ estim  : num  -0.0207 -0.1822 0.056 -0.1498 0.157 ...
#>  $ std.err: num  0.131 0.157 0.142 0.161 0.124 ...
#>  $ niter  : int  3 4 4 4 4 3 4 3 4 4 ...
#>  $ z.score: num  -0.158 -1.159 0.394 -0.931 1.268 ...
#>  $ p.value: num  0.874 0.246 0.693 0.352 0.205 ...

summary(glm(y01 ~ X1 + covar, family = "binomial"))$coefficients[2, ]

#>    Estimate  Std. Error     z value    Pr(>|z|) 
#> -0.02067582  0.13071424 -0.15817571  0.87431834 

# With only half of the data
ind.train <- sort(sample(n, n/2))
str(big_univLogReg(X.desc, y01[ind.train],
                   covar.train = covar[ind.train, ],
                   ind.train = ind.train))

#> 'data.frame':	4542 obs. of  5 variables:
#>  $ estim  : num  0.06151 -0.25252 -0.35347 -0.00754 -0.01936 ...
#>  $ std.err: num  0.192 0.222 0.218 0.233 0.172 ...
#>  $ niter  : int  4 4 4 3 3 4 4 4 4 4 ...
#>  $ z.score: num  0.3205 -1.1356 -1.6183 -0.0323 -0.1126 ...
#>  $ p.value: num  0.749 0.256 0.106 0.974 0.91 ...

summary(glm(y01 ~ X1 + covar, family = "binomial",
            subset = ind.train))$coefficients[2, ]

#>   Estimate Std. Error    z value   Pr(>|z|) 
#> 0.06150708 0.19193113 0.32046430 0.74861639