bigstatsr: Statistical Tools for Bigmemory Matrices

Easy-to-use, efficient, flexible and scalable statistical tools. Use bigmemory matrices.

Arguments

X

A big.matrix. You shouldn't have missing values in your data.

y.train

Vector of responses, corresponding to ind.train.

y01.train

Vector of responses, corresponding to ind.train. Must be 0s and 1s.

ind.train

An optional vector of the row indices that are used, for the training part. If not specified, all rows are used.

ind.row

An optional vector of the row indices that are used. If not specified, all rows are used.

ind.col

An optional vector of the column indices that are used. If not specified, all columns are used.

block.size

Maximum number of columns read at once. Default is 1000. This parameter controls the trade-off between memory usage and speed. Basically, the more you can load at once, the quicker will be the execution time, at the expense of memory usage.

ncores

Number of cores used. Default doesn't use parallelism.

ncores2

Number of cores used. Default doesn't use parallelism. For this function, use only half of the cores you have.

fun.scaling

A function that returns a named list of mean and sd for every column, to scale each of their elements such as followed: $$\frac{X_{i,j} - mean_j}{sd_j}$$.

covar.train

Matrix of covariables to be added in each model to correct for confounders (e.g. the scores of PCA), corresponding to ind.train. Default is NULL and corresponds to only adding an intercept to each model.

covar.row

Matrix of covariables to be added in each model to correct for confounders (e.g. the scores of PCA), corresponding to ind.row. Default is NULL and corresponds to only adding an intercept to each model.

use.Eigen

Should the Eigen library be used for matrix computations? Default tries to detect MRO. See details.

Details

For matrix computations, using Eigen library is faster. However, if you link R with an optimized math library, using R's base operations is even much faster.

For example, you can easily link R with the https://software.intel.com/en-us/intel-mkl (Intel® MKL) through https://mran.revolutionanalytics.com/open/ (MRO). It really improves performance of R and RcppArmadillo matrix computations, yet not the ones of RcppEigen (at least not directly).

So,

Eigen should be prefered if you don't change anything,

base R should be prefered if you use MRO,

Eigen may be prefered if you manage to link RcppEigen with the MKL (please mailto:florian.prive.21@gmail.com if you do!).

See also

Useful links:

https://privefl.github.io/bigstatsr

Report bugs at https://github.com/privefl/bigstatsr/issues