bigstatsr: Statistical Tools for Bigmemory Matrices

Easy-to-use, efficient, flexible and scalable statistical tools. Use bigmemory matrices.

Arguments

X

A big.matrix. You shouldn't have missing values in your data.

y

Vector of responses.

ind.train

An optional vector of the row indices that are used, for the training part. If not specified, all data are used.

block.size

Maximum number of columns read at once. Default is 1000.

ncores

Number or cores used. Default doesn't use parallelism.

fun.scaling

A function that returns a named list of mean and sd for every column, to scale each of their elements such as followed: \(\frac{X_{i,j} - mean_j}{sd_j}\).

use.Eigen

Should the Eigen library be used for matrix computations? Default is TRUE. See details.

Details

For matrix computations, using Eigen library is faster. However, if you link R with an optimized math library, using R's base operations is even much faster.

For example, you can easily link R with the https://software.intel.com/en-us/intel-mkl (Intel® MKL) through https://mran.revolutionanalytics.com/open/ (MRO). It really improves performance of R and RcppArmadillo matrix computations, yet not the ones of RcppEigen (at least not directly).

So,

Eigen should be prefered if you don't change anything,

base R should be prefered if you use MRO,

Eigen may be prefered if you manage to link RcppEigen with the MKL (please mailto:florian.prive.21@gmail.com if you do!).

See also

Useful links:

https://github.com/privefl/bigstatsr

Report bugs at https://github.com/privefl/bigstatsr/issues