Predict method

Predict method for class big_sp.

# S3 method for big_sp
predict(object, X, ind.row = rows_along(X),
  ind.col = object$ind.col, covar.row = NULL,
  block.size = block_size(nrow(X)), ...)

Arguments

object
Object of class big_sp.

X
A FBM.

ind.row
An optional vector of the row indices that are used. If not specified, all rows are used. Don't use negative indices.

ind.col
An optional vector of the column indices that are used. If not specified, all columns are used. Don't use negative indices.

covar.row
Matrix of covariables to be added in each model to correct for confounders (e.g. the scores of PCA), corresponding to ind.row. Default is NULL and corresponds to only adding an intercept to each model.

block.size
Maximum number of columns read at once. Default uses block_size.

...
Not used.

Value

A matrix of scores, with rows corresponding to ind.row and columns corresponding to lambda.

See also

big_spLinReg and big_spLogReg.