Remove temporary backing files

Remove temporary backing files

tmpFBM.rm(X.)

Arguments

X.
Either a big.matrix or a big.matrix.descriptor.

Value

0 for success, 1 for failure, invisibly. Not deleting a non-existent file is not a failure, nor is being unable to delete a directory if recursive = FALSE. However, missing values in x are regarded as failures.

Examples

list.files(path = tempdir())

#> [1] "36d0262954bb4cc0994c4a0a26ad07aa.png"                       
#> [2] "file4d512893c6a6.bk"                                        
#> [3] "file4d512893c6a6.desc"                                      
#> [4] "file4d5147a666a4.bk"                                        
#> [5] "file4d5147a666a4.desc"                                      
#> [6] "file4d515785c3a9.bk"                                        
#> [7] "file4d515785c3a9.desc"                                      
#> [8] "repos_https%3A%2F%2Fcran.rstudio.com%2F%2Fsrc%2Fcontrib.rds"
#> [9] "rs-graphics-b0800297-4574-42ae-8ab4-6c819f1c8dce"           

X.desc <- tmpFBM()(10, 10)
list.files(path = tempdir())

#>  [1] "36d0262954bb4cc0994c4a0a26ad07aa.png"                       
#>  [2] "file4d512893c6a6.bk"                                        
#>  [3] "file4d512893c6a6.desc"                                      
#>  [4] "file4d5147a666a4.bk"                                        
#>  [5] "file4d5147a666a4.desc"                                      
#>  [6] "file4d514e12d126.bk"                                        
#>  [7] "file4d514e12d126.desc"                                      
#>  [8] "file4d515785c3a9.bk"                                        
#>  [9] "file4d515785c3a9.desc"                                      
#> [10] "repos_https%3A%2F%2Fcran.rstudio.com%2F%2Fsrc%2Fcontrib.rds"
#> [11] "rs-graphics-b0800297-4574-42ae-8ab4-6c819f1c8dce"           

tmpFBM.rm(X.desc)
list.files(path = tempdir())

#> [1] "36d0262954bb4cc0994c4a0a26ad07aa.png"                       
#> [2] "file4d512893c6a6.bk"                                        
#> [3] "file4d512893c6a6.desc"                                      
#> [4] "file4d5147a666a4.bk"                                        
#> [5] "file4d5147a666a4.desc"                                      
#> [6] "file4d515785c3a9.bk"                                        
#> [7] "file4d515785c3a9.desc"                                      
#> [8] "repos_https%3A%2F%2Fcran.rstudio.com%2F%2Fsrc%2Fcontrib.rds"
#> [9] "rs-graphics-b0800297-4574-42ae-8ab4-6c819f1c8dce"