Remove temporary backing files

Remove temporary backing files

tmpFBM.rm(X.)

Arguments

X.
Either a big.matrix or a big.matrix.descriptor.

Value

These functions return a logical vector indicating which operation succeeded for each of the files attempted. Using a missing value for a file or path name will always be regarded as a failure.

If showWarnings = TRUE, file.create will give a warning for an unexpected failure.

Examples

list.files(path = tempdir())

#>  [1] "17afa3c7ecaf44eba43b019752943c95.png"                       
#>  [2] "216fd42cd6bb49d0b6ba7e984ae5116a.png"                       
#>  [3] "25870e91d7714b7181072d90ced2fe0c.png"                       
#>  [4] "3818ae1176e249dba5e646c08e2e078a.png"                       
#>  [5] "3c5450f090184a679610a53f487beb9b.png"                       
#>  [6] "45edacf523284c10875136192e15deba.png"                       
#>  [7] "5bd772ed8c064461896e0012955a38d8.png"                       
#>  [8] "80d2c71aa55d420688d3ee8d49f82f81.png"                       
#>  [9] "b0d9e6e5ef6a46dab167ea1009fc7620.png"                       
#> [10] "ed82d3fdd7c3404094715532ce9d1e75.png"                       
#> [11] "file17801af602d1.bk"                                        
#> [12] "file17801af602d1.desc"                                      
#> [13] "file17803ce5e25f.bk"                                        
#> [14] "file17803ce5e25f.desc"                                      
#> [15] "file17808febacd.bk"                                         
#> [16] "file17808febacd.desc"                                       
#> [17] "repos_https%3A%2F%2Fcran.rstudio.com%2F%2Fsrc%2Fcontrib.rds"

X.desc <- tmpFBM()(10, 10)
list.files(path = tempdir())

#>  [1] "17afa3c7ecaf44eba43b019752943c95.png"                       
#>  [2] "216fd42cd6bb49d0b6ba7e984ae5116a.png"                       
#>  [3] "25870e91d7714b7181072d90ced2fe0c.png"                       
#>  [4] "3818ae1176e249dba5e646c08e2e078a.png"                       
#>  [5] "3c5450f090184a679610a53f487beb9b.png"                       
#>  [6] "45edacf523284c10875136192e15deba.png"                       
#>  [7] "5bd772ed8c064461896e0012955a38d8.png"                       
#>  [8] "80d2c71aa55d420688d3ee8d49f82f81.png"                       
#>  [9] "b0d9e6e5ef6a46dab167ea1009fc7620.png"                       
#> [10] "ed82d3fdd7c3404094715532ce9d1e75.png"                       
#> [11] "file178011f1e4df.bk"                                        
#> [12] "file178011f1e4df.desc"                                      
#> [13] "file17801af602d1.bk"                                        
#> [14] "file17801af602d1.desc"                                      
#> [15] "file17803ce5e25f.bk"                                        
#> [16] "file17803ce5e25f.desc"                                      
#> [17] "file17808febacd.bk"                                         
#> [18] "file17808febacd.desc"                                       
#> [19] "repos_https%3A%2F%2Fcran.rstudio.com%2F%2Fsrc%2Fcontrib.rds"

tmpFBM.rm(X.desc)

#> [1] TRUE TRUE

list.files(path = tempdir())

#>  [1] "17afa3c7ecaf44eba43b019752943c95.png"                       
#>  [2] "216fd42cd6bb49d0b6ba7e984ae5116a.png"                       
#>  [3] "25870e91d7714b7181072d90ced2fe0c.png"                       
#>  [4] "3818ae1176e249dba5e646c08e2e078a.png"                       
#>  [5] "3c5450f090184a679610a53f487beb9b.png"                       
#>  [6] "45edacf523284c10875136192e15deba.png"                       
#>  [7] "5bd772ed8c064461896e0012955a38d8.png"                       
#>  [8] "80d2c71aa55d420688d3ee8d49f82f81.png"                       
#>  [9] "b0d9e6e5ef6a46dab167ea1009fc7620.png"                       
#> [10] "ed82d3fdd7c3404094715532ce9d1e75.png"                       
#> [11] "file17801af602d1.bk"                                        
#> [12] "file17801af602d1.desc"                                      
#> [13] "file17803ce5e25f.bk"                                        
#> [14] "file17803ce5e25f.desc"                                      
#> [15] "file17808febacd.bk"                                         
#> [16] "file17808febacd.desc"                                       
#> [17] "repos_https%3A%2F%2Fcran.rstudio.com%2F%2Fsrc%2Fcontrib.rds"