Remove temporary backing files

Remove temporary backing files

tmpFBM.rm(X.)

Arguments

X.
Either a big.matrix or a big.matrix.descriptor.

Value

These functions return a logical vector indicating which operation succeeded for each of the files attempted. Using a missing value for a file or path name will always be regarded as a failure.

If showWarnings = TRUE, file.create will give a warning for an unexpected failure.

Examples

list.files(path = tempdir())

#>  [1] "1f12dc3d6d5c4acf84e94e002f27260a.png"                       
#>  [2] "25d19b4edb0e43cbbf0bea1bfb896520.png"                       
#>  [3] "28d5206c9a1942459aa790fefa1892f0.png"                       
#>  [4] "2ae76331a43e4bc9bb32f435e72564dc.png"                       
#>  [5] "772870fc351e439394f7dcbf0724c8ce.png"                       
#>  [6] "96d194cb4815439bbf8fd8b3844b902e.png"                       
#>  [7] "d1c95c1c122f49eba76e33843e6fa379.png"                       
#>  [8] "devtools_install_1a0b32ae1f8"                               
#>  [9] "ed48565628144017847e4fc1c6c16cb4.png"                       
#> [10] "file1a0b284e3955.bk"                                        
#> [11] "file1a0b284e3955.desc"                                      
#> [12] "file1a0b6b1eb5d5.bk"                                        
#> [13] "file1a0b6b1eb5d5.desc"                                      
#> [14] "file1a0b7b039e95.bk"                                        
#> [15] "file1a0b7b039e95.desc"                                      
#> [16] "gitmsg1a0b436d5d5.txt"                                      
#> [17] "repos_https%3A%2F%2Fcran.rstudio.com%2F%2Fsrc%2Fcontrib.rds"
#> [18] "Rprofile-devtools"                                          

X.desc <- tmpFBM()(10, 10)
list.files(path = tempdir())

#>  [1] "1f12dc3d6d5c4acf84e94e002f27260a.png"                       
#>  [2] "25d19b4edb0e43cbbf0bea1bfb896520.png"                       
#>  [3] "28d5206c9a1942459aa790fefa1892f0.png"                       
#>  [4] "2ae76331a43e4bc9bb32f435e72564dc.png"                       
#>  [5] "772870fc351e439394f7dcbf0724c8ce.png"                       
#>  [6] "96d194cb4815439bbf8fd8b3844b902e.png"                       
#>  [7] "d1c95c1c122f49eba76e33843e6fa379.png"                       
#>  [8] "devtools_install_1a0b32ae1f8"                               
#>  [9] "ed48565628144017847e4fc1c6c16cb4.png"                       
#> [10] "file1a0b284e3955.bk"                                        
#> [11] "file1a0b284e3955.desc"                                      
#> [12] "file1a0b44cdda28.bk"                                        
#> [13] "file1a0b44cdda28.desc"                                      
#> [14] "file1a0b6b1eb5d5.bk"                                        
#> [15] "file1a0b6b1eb5d5.desc"                                      
#> [16] "file1a0b7b039e95.bk"                                        
#> [17] "file1a0b7b039e95.desc"                                      
#> [18] "gitmsg1a0b436d5d5.txt"                                      
#> [19] "repos_https%3A%2F%2Fcran.rstudio.com%2F%2Fsrc%2Fcontrib.rds"
#> [20] "Rprofile-devtools"                                          

tmpFBM.rm(X.desc)

#> [1] TRUE TRUE

list.files(path = tempdir())

#>  [1] "1f12dc3d6d5c4acf84e94e002f27260a.png"                       
#>  [2] "25d19b4edb0e43cbbf0bea1bfb896520.png"                       
#>  [3] "28d5206c9a1942459aa790fefa1892f0.png"                       
#>  [4] "2ae76331a43e4bc9bb32f435e72564dc.png"                       
#>  [5] "772870fc351e439394f7dcbf0724c8ce.png"                       
#>  [6] "96d194cb4815439bbf8fd8b3844b902e.png"                       
#>  [7] "d1c95c1c122f49eba76e33843e6fa379.png"                       
#>  [8] "devtools_install_1a0b32ae1f8"                               
#>  [9] "ed48565628144017847e4fc1c6c16cb4.png"                       
#> [10] "file1a0b284e3955.bk"                                        
#> [11] "file1a0b284e3955.desc"                                      
#> [12] "file1a0b6b1eb5d5.bk"                                        
#> [13] "file1a0b6b1eb5d5.desc"                                      
#> [14] "file1a0b7b039e95.bk"                                        
#> [15] "file1a0b7b039e95.desc"                                      
#> [16] "gitmsg1a0b436d5d5.txt"                                      
#> [17] "repos_https%3A%2F%2Fcran.rstudio.com%2F%2Fsrc%2Fcontrib.rds"
#> [18] "Rprofile-devtools"